% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{add.covariance.of.estimable.unit.estimable.model}
\alias{add.covariance.of.estimable.unit.estimable.model}
\title{add covariance relations for inner random variable of \code{estimable.unit}s in an \code{estimable.model}}
\usage{
add.covariance.of.estimable.unit.estimable.model(esm, v1, v2, x,
  v1.subind = 1, v2.subind = 1, do.check = TRUE)
}
\arguments{
\item{esm}{estimable model}

\item{v1}{estimable.unit 1}

\item{v2}{estimable.unit 2}

\item{x}{a new value, should be one number}

\item{v1.subind}{one estimable.unit may have more than one random varible in it, this indicate which variable are operated on}

\item{v2.subind}{similar to \code{v1.subind}}

\item{do.check}{check if the input \code{v1,v2,v1.subind,v2.subind} are valid}
}
\value{
the updated estimable.model
}
\description{
add covariance relations for inner random variable of \code{estimable.unit}s in an \code{estimable.model}
}
\note{
This function is a scaler version, all arguments should be scalar.
}
