% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.003eqns.R
\name{analyse.0.0}
\alias{analyse.0.0}
\title{analyse.0.0}
\usage{
analyse.0.0(eqns, ...)
}
\arguments{
\item{eqns}{equations}

\item{...}{extra.rules}
}
\value{
the found definitions for pk variables, and other non-pk definitions are kept as is
}
\description{
After dividing up the graph into connected.components, 
we do \code{\link{analyse.connected.component.critical.points}} on each one for extracting pk variables definitions.
for other unknown variables, currently we just keep them as is.
We may add more code for analysing those part(the other part) in future.
}
