% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30pattern.001analyseConnected.R
\name{analyse.0.1}
\alias{analyse.0.1}
\title{find parameter spaces embedding in the result of \code{analyse.0.0}}
\usage{
analyse.0.1(cps)
}
\arguments{
\item{cps}{connected components, the \code{analysed.defns} of returned object of \code{analyse.0.0}}
}
\value{
if any duplication(one parameter used by multiple critical variables) is found
}
\description{
A possible solution for this issue, is to add one more layer as the embedding from a lower dimensional
space into the full dimensional parameter space
}
