% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30pattern.001analyseConnected.R
\name{analyse.common.used.parameter.in.one.component}
\alias{analyse.common.used.parameter.in.one.component}
\title{analyse one connected component
to check if there is any following patterns}
\usage{
analyse.common.used.parameter.in.one.component(cp)
}
\arguments{
\item{cp}{connected.component}
}
\value{
found parameters used by multiple critical parameters
}
\description{
analyse one connected component
to check if there is any following patterns
}
