% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.003eqns.R
\name{analyse.connected.component.first.critical.points}
\alias{analyse.connected.component.first.critical.points}
\title{analyse.connected.component.critical.points}
\usage{
analyse.connected.component.first.critical.points(eqns, socp,
  extra.rules = NULL, skip.patterns = FALSE, ...)
}
\arguments{
\item{eqns}{equations}

\item{socp}{set of critical points, critical points should be an lhs, and is necessary for its offsprins critical points}

\item{extra.rules}{also allow to add a list of pattern and actions to extend the functionality of this function}

\item{skip.patterns}{if skip pattern matchers}

\item{...}{other parameters may be passed to the pattern functions}
}
\value{
information on variability
}
\description{
The working horse for \code{analyse.connected.component.critical.points}
combind \code{topo.sort},
\code{get.definition.of.v},
and use \code{pattern.distrib.*}
to analysis to some extent general equations
assuming eqns are connected.component
Try to figure out the meaning of the critical points in the connected component
And return the left ind of the eqns
if no PK lhs is found, return just the left inds
}
