% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.001analyse.R
\name{analyse.group.effect}
\alias{analyse.group.effect}
\title{analyse.group.effect}
\usage{
analyse.group.effect(e)
}
\arguments{
\item{e}{expression}
}
\value{
if the expression has a known pattern as multiplicative, additive, affine and exponential
}
\description{
for a expression, determine if there is a recognizable structure of the mean expression
}
\details{
the slot \code{parameter.space} in return object is for back referring to the \code{THETA[i]}
what can be used for look up the definition domain of the parameters
}
