% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.080morphism.domain.R
\name{as.low.upper.list.morphism.domain}
\alias{as.low.upper.list.morphism.domain}
\title{convert morphism
to a list of upper and lower ranges of the parameters}
\usage{
as.low.upper.list.morphism.domain(m)
}
\arguments{
\item{m}{morphism}
}
\value{
list
}
\description{
convert morphism
to a list of upper and lower ranges of the parameters
}
