% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.002dynamic.system.R
\name{as.nonmem.subroutine.dynamics}
\alias{as.nonmem.subroutine.dynamics}
\title{convert description in kinetic to NONMEM PREDPP subroutine}
\usage{
as.nonmem.subroutine.dynamics(d)
}
\arguments{
\item{d}{kinetic description}
}
\value{
string used in NONMEM control file SUBR block
}
\description{
convert description in kinetic to NONMEM PREDPP subroutine
}
