% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{as.symbolic.vector}
\alias{as.symbolic.vector}
\title{enforce a input symbolic element to be a list}
\usage{
as.symbolic.vector(e)
}
\arguments{
\item{e}{expression}
}
\value{
a list if success
}
\description{
enforce a input symbolic element to be a list
}
