% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{asCharacterSymbolCall}
\alias{asCharacterSymbolCall}
\title{Change a symbol like CL or a call like A[1] to a character}
\usage{
asCharacterSymbolCall(x, keep.original = FALSE)
}
\arguments{
\item{x}{symbol or call}

\item{keep.original}{if save the original call/symbol in attribute}
}
\value{
the character
}
\description{
Change a symbol like CL or a call like A[1] to a character
}
