% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{asCMatrix.RMatrix}
\alias{asCMatrix.RMatrix}
\title{convertion, listoflist to matrix}
\usage{
asCMatrix.RMatrix(A)
}
\arguments{
\item{A}{RMatrix}
}
\value{
CMatrix
}
\description{
convertion, listoflist to matrix
}
