% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{asRMatrix.CMatrix}
\alias{asRMatrix.CMatrix}
\title{convert list to listoflist matrix}
\usage{
asRMatrix.CMatrix(A)
}
\arguments{
\item{A}{CMatrix}
}
\value{
RMatrix
}
\description{
convert list to listoflist matrix
}
