% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.080morphism.domain.R
\name{backref.parameter.of.morphism}
\alias{backref.parameter.of.morphism}
\title{referring to original THETA's
We store the \code{THETA[i]}'s index in \code{m$defn$parameter.space}
or \code{m$defn$defn$parameter.space} if \code{m} is a \code{groupeffect} type
when we need a value or range by lookup from original nmModel
this pointer is needed}
\usage{
backref.parameter.of.morphism(m)
}
\value{
integer vector for backreferring to original index of THETAs
}
\description{
referring to original THETA's
We store the \code{THETA[i]}'s index in \code{m$defn$parameter.space}
or \code{m$defn$defn$parameter.space} if \code{m} is a \code{groupeffect} type
when we need a value or range by lookup from original nmModel
this pointer is needed
}
