% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.080morphism.domain.R
\name{calculate.ambience.domain}
\alias{calculate.ambience.domain}
\title{A^1 * B^2 * C^3 => ambienceDomain^6
Which means, the domain A^1 * B^2 * C^3 has to be embeded into a six dimensional Euclidean space
Of course, we assume baseDomains should be real, not complex numbers}
\usage{
calculate.ambience.domain(e)
}
\description{
A^1 * B^2 * C^3 => ambienceDomain^6
Which means, the domain A^1 * B^2 * C^3 has to be embeded into a six dimensional Euclidean space
Of course, we assume baseDomains should be real, not complex numbers
}
