% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{character.rules.to.list.rules}
\alias{character.rules.to.list.rules}
\title{character.rules.to.list.rules}
\usage{
character.rules.to.list.rules(s0)
}
\arguments{
\item{s0}{a character vector}
}
\value{
list of pairs of expressions
}
\description{
a syntaxtic surger for translate character written rules into list of list form
}
\details{
Character vector containing rules should look like following: \cr
 \code{
     c( lhs1, rhs1,
        lhs2, rhs2, 
        ... , ... ,
      ) } \cr
And those strings are parsed and paired into list of pair of expressions
}
\seealso{
\code{\link{symbolic.simplify.gigo}}
}
\author{
jjxie
}
