% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{check.compatibility.of.covariance.structure}
\alias{check.compatibility.of.covariance.structure}
\title{Check if the additional convariance is compatible with the model definition}
\usage{
check.compatibility.of.covariance.structure(esm, need.print = TRUE)
}
\arguments{
\item{esm}{estimable.model}

\item{need.print}{if print diagonize information}
}
\value{
vector of TRUE or FALSE indication if compatible for PK.COV or OBS.COV
}
\description{
We allow user to change the estimable model,
usually, the distribution type or covariance matrix of parameters of \code{estimable.model} can be changed.
Even some \code{estimable.unit}s or \code{random.observation}s can be added into or removed from the \code{estimable.model}.
The change may cause conflications between the \code{PK.PROGRAM} or \code{OBS.PROGRAM} and
 the additional convariance structures defined in \code{PK.COV} or \code{OBS.COV}.
This function will check the situation and print a warning.
}
