% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.110morphism.combinators.R
\name{check.groupeffect.definition}
\alias{check.groupeffect.definition}
\title{just a check for groupeffect definition, and throw a exception if necessary
any morphism, or group definition should be THETA(parameters) free}
\usage{
check.groupeffect.definition(gpdefn = NULL)
}
\arguments{
\item{gpdefn}{\code{grpvar.definition}}
}
\value{
TRUE or simplify throw exceptions
}
\description{
just a check for groupeffect definition, and throw a exception if necessary
any morphism, or group definition should be THETA(parameters) free
}
