% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{check.isvalid.basic.pk.parameters}
\alias{check.isvalid.basic.pk.parameters}
\title{check if all basic pk parameter needed by subroutine and trans appearing in the PK.PROGRAM}
\usage{
check.isvalid.basic.pk.parameters(esm, need.print = TRUE)
}
\arguments{
\item{esm}{estimable.model}

\item{need.print}{print diagonize information}
}
\value{
TRUE or FALSE
}
\description{
check if all basic pk parameter needed by subroutine and trans appearing in the PK.PROGRAM
}
