% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{check.meta.information}
\alias{check.meta.information}
\title{Check if the meta info is compatible with the \code{PK.PROGRAM} or \code{OBS.PROGRAM} code in the model}
\usage{
check.meta.information(esm, need.print = TRUE)
}
\arguments{
\item{esm}{estimable.model}

\item{need.print}{print diagonize information}
}
\value{
TRUE or FALSE
}
\description{
The story is, after deleting some \code{estimable.unit} which is essential for the model definition, 
it will cause problem. We will check before convert it into a control file.
}
