% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{collect.algebraical.1}
\alias{collect.algebraical.1}
\title{collect.algebraical.1}
\usage{
collect.algebraical.1(e)
}
\arguments{
\item{e}{expression}
}
\value{
expression
}
\description{
\code{ a*a*a + b*b --> a^3 + b^2 } \cr
\code{ (a+b)*(a+b) --> a^2 + 2*a*b + b^2 }
}
\author{
jjxie
}
