% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{collect.atomic.product.list}
\alias{collect.atomic.product.list}
\title{collect.atomic.product.list}
\usage{
collect.atomic.product.list(le)
}
\arguments{
\item{le}{list of expression}
}
\value{
list of expression which atomics are combined
}
\description{
compute and combine atomics in a list of expression representing a production\cr
\code{[ A, B, 3, 4, -1 ] -> [ -12, A, B ]}
}
\author{
jjxie
}
