% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{collect.atomic.sum.list}
\alias{collect.atomic.sum.list}
\title{collect.atomic.sum.list}
\usage{
collect.atomic.sum.list(le)
}
\arguments{
\item{le}{list of expression}
}
\value{
list of expression which atomics are combined
}
\description{
compute and combine list of expressions representing a summation, similar to \code{\link{collect.atomic.product.list}} \cr
\code{[ A, B, 3, 4, -1 ] -> [ 6, A, B ]}
}
\author{
jjxie
}
