% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25plainmath.001matrix.R
\name{collect.matrix.list.to.diagonal.and.block}
\alias{collect.matrix.list.to.diagonal.and.block}
\title{Give more concise description of matrix list
assume each entry in ml are square}
\usage{
collect.matrix.list.to.diagonal.and.block(ml, prefix = "ETA")
}
\arguments{
\item{ml}{matrix list}

\item{prefix}{prefix for constructing names}
}
\value{
list of diagonal and block
}
\description{
Give more concise description of matrix list
assume each entry in ml are square
}
