% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.110morphism.combinators.R
\name{combine.affinemorphism.factor}
\alias{combine.affinemorphism.factor}
\title{combine a affine mapping and a factor (with same number of levels)
generate a affine.mapping.factor.effect}
\usage{
combine.affinemorphism.factor(m, var.name, levels)
}
\arguments{
\item{var.name}{name}
}
\value{
morphism
}
\description{
combine a affine mapping and a factor (with same number of levels)
generate a affine.mapping.factor.effect
}
