% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{compare.estimable.model}
\alias{compare.estimable.model}
\title{compare two estimable model}
\usage{
compare.estimable.model(m1, m2, model.names = c("Model1", "Model2"))
}
\arguments{
\item{m1}{model spec 1}

\item{m2}{model spec 2}

\item{model.names}{names used in output table}
}
\value{
summary table
}
\description{
The result is a table showing the type and typical value and initial estimations of basic PK parameters.
}
\details{
If two model are of different kinetic model (using different PREDPP subroutine), we do not real compare them, but output the subroutine they used.
}
