% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{compare.estimable.unit}
\alias{compare.estimable.unit}
\title{compare two estimable units}
\usage{
compare.estimable.unit(m1, m2)
}
\arguments{
\item{m1}{first  estimable unit}

\item{m2}{second estimable unit}
}
\value{
data.frame
}
\description{
compare two estimable units
}
