% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\name{compile.ifelse.to.if.statements}
\alias{compile.ifelse.to.if.statements}
\title{Compile \code{ifelse} function call to \code{if} statements}
\usage{
compile.ifelse.to.if.statements(e0, simplify.result = TRUE,
  remove.na.init = TRUE, temp.var.prefix = "TMPVAR")
}
\arguments{
\item{simplify.result}{if call \code{eliminate.variable.alias.in.eqns}}

\item{remove.na.init}{if call \code{eliminate.na.initilization.in.eqns}}

\item{temp.var.prefix}{the prefix of temporary varibles used in output code}

\item{e}{expression}
}
\value{
equations which are equivlent to e, using non-nested IF/ELSE
}
\description{
Compile \code{ifelse} function call to \code{if} statements
}
