% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.003estimable.model.R
\name{compile.model.spec.to.nonmem.spec}
\alias{compile.model.spec.to.nonmem.spec}
\title{Compiler which compiles \code{estimable.model} into a NONMEM like specification}
\usage{
compile.model.spec.to.nonmem.spec(ms)
}
\arguments{
\item{ms}{model spec by \code{create.model.spec.from.nmmod.and.nmdata}}
}
\value{
a NONMEM specification, which should be quite familar to NONMEM modelers
}
\description{
This is the main working horse of the exporting phase.
}
