% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.regressors.names.R
\name{compute.regressors.names}
\alias{compute.regressors.names}
\title{Compute all possible formula terms from a given dataset}
\usage{
compute.regressors.names(complete.X.df, n.squares, transformations)
}
\arguments{
\item{complete.X.df}{The dataset that contains the base variables the formula is composed of (column-wise)}

\item{n.squares}{The maximum order of the polynomial composition of base variables. Ex. \verb{order 0 = a}, \verb{order 1 = a*b}, \verb{order 2 = a*b*c}}

\item{transformations}{A list of potentially non-linear transformations that can be applied on top of the squares. Ex. \verb{order 0, transformation=log10 = log10.a}. Input values are x and z, the array of numbers to be transformed (training set only), and the min max statistics (on the global dataset) respectively.}
}
\description{
Compute all possible formula terms from a given dataset
}
\examples{
\dontrun{
 complete.regressors <- compute.regressors.names(complete.X.df, n.squares, transformations)
 print(length(complete.regressors))
}
}
