% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{compute.symbolic}
\alias{compute.symbolic}
\title{compute.symbolic
This is example for compute the symbolic value given several equations\cr
implementation is easy, because we have evaluation engine already, we can even \cr
write program in that "language" \cr
here we just need to evaluate the symbol in the populated environment}
\usage{
compute.symbolic(eqns, v)
}
\arguments{
\item{eqns}{equations, nonmem characters or just R's expressions}
}
\value{
list, with the name
}
