% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05graphic.000common.R
\name{connected.component}
\alias{connected.component}
\title{get the connected component}
\usage{
connected.component(G, v, direction = 0)
}
\arguments{
\item{G}{graph}

\item{v}{vertices set from which looking}

\item{direction}{0 denotes incoming and 1 denotes outcoming, and c(0,1) for both directions}
}
\value{
set of vertex
}
\description{
connected.component
}
\author{
jjxie
}
