% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.070constructor.R
\name{construct.obs.variable}
\alias{construct.obs.variable}
\title{construct.obs.variable
similar to \code{\link{construct.typical.variable}}}
\usage{
construct.obs.variable(type, last.ind = 0, obs.varname, defn)
}
\arguments{
\item{type}{type of observation variable}

\item{last.ind}{last used \code{EPS} index}

\item{obs.varname}{the name for observation variable}

\item{defn}{definition in the errorObject}
}
\description{
construct.obs.variable
similar to \code{\link{construct.typical.variable}}
}
