% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.070constructor.R
\name{construct.random.variable}
\alias{construct.random.variable}
\title{construct.random.variable
similar to \code{\link{construct.typical.variable}}}
\usage{
construct.random.variable(distrib, Var, pkvar, typical.var, last.ind = 0,
  random.var = quote(ETA), exp.mean.var = NULL)
}
\arguments{
\item{distrib}{distribution for the PK variable}

\item{Var}{variation for the PK variable}

\item{pkvar}{pk variable name}

\item{typical.var}{typical variable name}

\item{last.ind}{last used \code{ETA} index}

\item{random.var}{random symbol}

\item{exp.mean.var}{we can use \code{exp.mean.var} to override the construct of it from \code{typical.var}}
}
\value{
construct with expression and index information
}
\description{
construct.random.variable
similar to \code{\link{construct.typical.variable}}
}
