% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.070constructor.R
\name{construct.typical.variable}
\alias{construct.typical.variable}
\title{typical values constructor}
\usage{
construct.typical.variable(grptype, typical.var, last.ind = 0, defn, pkvar)
}
\arguments{
\item{grptype}{group effect}

\item{typical.var}{typical value name}

\item{last.ind}{the last used index for \code{THETA}}

\item{defn}{definition for the target pk variable}

\item{pkvar}{pk variable name}
}
\description{
typical values constructor
}
