% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{convert.distrib.description.to.random.disturbation}
\alias{convert.distrib.description.to.random.disturbation}
\title{convertion}
\usage{
convert.distrib.description.to.random.disturbation(type, Var = 1)
}
\arguments{
\item{type}{distribution type}

\item{Var}{if atomic, assuming just a RANDOM type, or variation expression \code{coeff*Var(ETA[i])}, then \code{sqrt(coeff)}}
}
\value{
a random.disturbation object
}
\description{
convertion
}
