% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{convert.matrix.to.affine.mapping}
\alias{convert.matrix.to.affine.mapping}
\title{convert atomic matrix to symbolic affine mapping}
\usage{
convert.matrix.to.affine.mapping(A, b = NULL)
}
\arguments{
\item{A}{matrix}

\item{b}{optional vector}
}
\value{
affine.mapping
}
\description{
convert atomic matrix to symbolic affine mapping
}
