% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.110morphism.combinators.R
\name{convert.matrix.to.morphism}
\alias{convert.matrix.to.morphism}
\title{atomic matrix to affine.mapping morphism}
\usage{
convert.matrix.to.morphism(A, b = NULL)
}
\arguments{
\item{A}{matrix}

\item{b}{optional vector}
}
\value{
AffineMapping morphism
}
\description{
atomic matrix to affine.mapping morphism
}
