% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{Cov.random.disturbation}
\alias{Cov.random.disturbation}
\title{The covariance of the random part of a disturbation}
\usage{
Cov.random.disturbation(rd)
}
\arguments{
\item{rd}{random distribution}
}
\value{
the covariance matrix
}
\description{
The covariance of the random part of a disturbation
}
