% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{create.abstractGraph.from.equations}
\alias{create.abstractGraph.from.equations}
\title{create.abstractGraph.from.equations}
\usage{
create.abstractGraph.from.equations(eqns)
}
\arguments{
\item{eqns}{equations expressions}
}
\description{
Note: Each equation should be in a single assignment form as: \code{LHS = f(RHS1,RHS2,...)}
This is not correct for a general statement. e.g. if(..) { a1=f1(...);a2=f2(...) } else { a1=f1'(...); a2=f2'(...) } 
It's not correct to construct the full binary graph (i.e. connect each lhs and rhs vertex), a new code is added to handle this.
}
\details{
And , duplicated edges will be generated, if you put two same lines there:
\code{   A = B
         A = B }
will give two same edges in graph, that from B pointing to A.

similar to \code{\link{create.graph.from.equations}}
but in edge representation, the vertex are not characters, just numbers,
moreover, we record the vertices table of expressions
}
