% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{create.firstorder.linear.ode}
\alias{create.firstorder.linear.ode}
\title{create.firstorder.linear.ode}
\usage{
create.firstorder.linear.ode(state.vector, A)
}
\arguments{
\item{state.vector}{a list of symbols or \code{TUPLE(x1,x2,x3)} like expression which will be converted into list automatically}

\item{A}{matrix should be list of list type}
}
\value{
ODE object
}
\description{
ODE only consider very simple ones as following
\code{ dx/dt = A \%*\% x }
}
