% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{create.graph.from.equations}
\alias{create.graph.from.equations}
\title{create.graph.from.equations}
\usage{
create.graph.from.equations(lines)
}
\arguments{
\item{lines}{text lines from NONMEM control stream, might be PK, PRED, ERROR, DES and e.t.c. blocks}
}
\value{
list where \code{eqns.ori} parsed equations,\cr
 \code{eqns.lhs} and \code{eqns.rhs} are left and right hand side of equations, \cr
 \code{eqns.rel} simplified equations, and tranformed using \code{\link{nonmem.parameters.specialarray.to.r.array}} \cr
 \code{edges.ref.eqn} back reference from indices of edges to indices of equations in \code{eqns.rel}, \cr
 \code{graph} the graph
}
\description{
For  system of equations, create dependency equations\cr
For  \code{A = f(X1,X2,X3,...XN)} \cr
  create edges \code{<X1,A>, <X2,A> , ... , <XN,A>} \cr
The whold equations will genearte a Graph.
}
\details{
However, there might be more than one \code{<Xi,Yi>} edges in the graph, e.g.  \cr
\code{ CL = TVCL * ETA(1) } \cr
\code{ CL = TVCL * ETA(2) } \cr
will generate two \code{<TVCL,CL>} edges, hence the index is important \cr
(of course, we do not need to include it explicitly in the return object) \cr
when restrict to subgraph, we need only morphism (see \code{\link{restriction.morphism.of.graph.at.v}})
}
\author{
Mango solutions
}
