% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.003estimable.model.R
\name{create.model.spec.from.nmmod.and.nmdata}
\alias{create.model.spec.from.nmmod.and.nmdata}
\title{From nmmod and nmdata (object returned by \code{RNMImport} package), generate the \code{estimable.model} (Model Spec)}
\usage{
create.model.spec.from.nmmod.and.nmdata(nmmod, modData, keep_mu_i_s = TRUE,
  try_abbr_pk_names = TRUE)
}
\arguments{
\item{nmmod}{the nonmem model return by \code{importNmMod}}

\item{modData}{the Input Data returned by \code{RNMImport:::importModelData}}

\item{keep_mu_i_s}{should keep MU_i in NONMEM statements}

\item{try_abbr_pk_names}{if try to catch PK parameters using abbreviation like ALPHA -> ALPH}
}
\value{
specification
}
\description{
This is the main API in this package behaving as a user level API.
It's working as an extension of the \code{RNMImport} pacakge. Mango's \code{RNMImport} deal with control file, or NONMEM run,
but the PK block(code) and ERROR block(code) in the control file are stored faithfully as plain strings. 
This function tries to do deeper information extraction from the return object of \code{RNMImport}.
This second phase analysing depends on the following assumptions.
Quite a few NONMEM control file are written using the pre-defined PREDPP compartment subroutines, and those subroutines have
definite input PK parameters, like \code{CL}, \code{V} and e.t.c. The variant part are how these parameters are constituted,
e.g. you may simplely use one parameter for typical value and additional populational random effect. But also you can model 
it with multiple parameters and also invole other covariables. 
So, even one simple kinetic model as ADVAN2 can derive tens of different concrete models. But there are something in common,
and something can be re-used or managed by computer, i.e. the constitution rules. We use \code{estimable.unit} and \code{random.observation} for them.
For the math, the \code{estimable.unit} can be a \code{morphism} or \code{random.disturbation}.
And based on above, the two models with same kinetic system can be compared, even they have different numbers of \code{THETA}s.
}
\details{
Of course, base on above assumptions, we do not support those NONMEM code written in \code{PRED} and those using a generic ODE kinetic system like
ADVAN6, e.t.c
}
\note{
Currently, this function resolve duplicated use of parameters by introducing new variables

\code{
 V = THETA[1]
 CL = THETA[1] + THETA[2] }
Then it will think it as
\code{
 THETAHAT01 = THETA[1]
 V = THETAHAT01
 CL = THETAHAT01 + THETA[2]}

Also, we assume the \code{nmmod} should have both PK and ERROR part.
If the control file, or say, the \code{nmmod} only has PRED code, then it hard to determine what the model
means. Now, we stop if only PRED appears.
}
