% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.002monomial.R
\name{create.polynomial.of.exp.of.random.variable}
\alias{create.polynomial.of.exp.of.random.variable}
\title{create.polynomial.of.exp.of.random.variable}
\usage{
create.polynomial.of.exp.of.random.variable(e)
}
\arguments{
\item{e}{expression}
}
\value{
list of list \code{[[coeff, monomial]]}
}
\description{
similar to \code{\link{create.polynomial.of.random.variable}} \cr
while monomials here are of form \code{exp(ETA[i])^di} \cr
\code{e --> [ [ coeff, main.monomial] ]}
}
\author{
Mango solutions
}
