% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.002monomial.R
\name{create.polynomial.of.random.variable}
\alias{create.polynomial.of.random.variable}
\title{create.polynomial.of.random.variable}
\usage{
create.polynomial.of.random.variable(e)
}
\arguments{
\item{e}{expression}
}
\value{
list of list representing sum of list like \code{ [ [coeff1, EPS[1]^1], [coeff2, EPS[1]^2] ] }
}
\description{
Convert an expression \code{e} to a polynomial of \code{ETA[i]}'s or \code{EPS[i]}'s as following:\cr
\code{ e --> [ [ coeff, main.monomial] ] }
}
\author{
Mango solutions
}
