% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.002monomial.R
\name{create.polynomial.of.theta}
\alias{create.polynomial.of.theta}
\title{create.polynomial.of.theta}
\usage{
create.polynomial.of.theta(e)
}
\arguments{
\item{e}{expression}
}
\value{
list of list as \code{ [ [coeff, main.monomial] ,...] }
}
\description{
similar as \code{\link{create.polynomial.of.random.variable}} \cr
\code{ e --> [ [ coeff, main.monomial] ] } \cr
Only difference is monomial here are of type \code{ THETA[i]^di }
}
\author{
Mango solutions
}
