% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.validate.R
\name{cross.validate}
\alias{cross.validate}
\title{Test a non-linear formula}
\usage{
cross.validate(
  cur.dataset,
  y,
  cur.vars,
  custom.abs.mins,
  K,
  N,
  transformations,
  cv.norm
)
}
\arguments{
\item{cur.dataset}{The dataset to be used for the test}

\item{y}{The independent variable}

\item{cur.vars}{An array of non-linear formula terms to be tested. \code{cur.vars <- c('a','mul.a.b')} will test the formula \code{y ~ a + a*b}}

\item{custom.abs.mins}{A list of user-defined minimum values for dataset columns.}

\item{K}{The number of parts the dataset is split into for K-fold cross-validation.}

\item{N}{The number of times the K-fold validation is repeated, shuffling the dataset row orders before each time.}

\item{transformations}{A list of potentially non-linear transformations allowed in \code{cur.vars}.}

\item{cv.norm}{Normalize regressors after train-validation split in inner cross-validation loop.}
}
\value{
A data.frame with cross-validation results for each fold (K) and for each round (N).
}
\description{
Test using K-fold cross-validation repeated N times
}
\examples{
\donttest{
# set-up a toy example dataset
x1<-runif(100, min=2, max=67)
x2<-runif(100, min=0.01, max=0.1)
X <- data.frame(x1=x1, x2=x2)
# set up a "true" non-linear relationship
# with some noise
y <- log10(x1^2*x2) + rnorm(100, 0, 0.001)

# do actual cross-validation experiments, record in experiments data.frame
# all the validation-set performances.
experiments.0 <- cross.validate(
    X, y,
    cur.vars=c('inv.x1','x2'),
    custom.abs.mins=list(),
    K=2,
    N=3,
   transformations=list(
      "log"=function(rdf, x, stats){ log(x) },
      "log_x1_p"=function(rdf, x, stats){ log(rdf$x1 + x) },
      "inv"=function(rdf, x, stats){ 1/x }
    ),
    cv.norm=FALSE
)

# summarize cross-validation results by averaging
errs.m <- stats::aggregate(
    cbind(base.pe, base.cor, base.r.squared,
          base.max.pe, base.iqr.pe, base.max.cooksd)~1,
   data=experiments.0, FUN=mean
)
}

}
