% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.min.maxs.R
\name{dataset.min.maxs}
\alias{dataset.min.maxs}
\title{Compute dataset column-wise statistics: min, absmin, absmax, projzero}
\usage{
dataset.min.maxs(X.df.std, X.mean.sd)
}
\arguments{
\item{X.df.std}{input dataset}

\item{X.mean.sd}{pre-computed column-wise statistics for the dataset, mean and sd}
}
\value{
A list of statistics for each column.

min: the minimum of the column values
absmin: the minimum of the absolute values of the columns
absmax: the maximum of the absolute values of the columns
projzero: -mean/sd of the columns, that is the position of the zero in the original, non-normalized space.
}
\description{
Compute dataset column-wise statistics: min, absmin, absmax, projzero
}
