% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{degree.of.freedom}
\alias{degree.of.freedom}
\title{extract the dimension of domain}
\usage{
degree.of.freedom(m)
}
\arguments{
\item{m}{morphism}
}
\value{
dimension number
}
\description{
extract the dimension of domain
}
