% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{design.matrix.of.one.factor}
\alias{design.matrix.of.one.factor}
\title{Of type, we deal with one variable only
\code{ifelse(var1==a1, .. , ifelse( var1==a2, .. ,..))}}
\usage{
design.matrix.of.one.factor(e, vn, vt = NULL, lvs = levels(vt))
}
\arguments{
\item{e}{expression}

\item{vt}{factor type variable}

\item{lvs}{also can give directly the levels}
}
\value{
design matrix of \code{vt}
}
\description{
Of type, we deal with one variable only
\code{ifelse(var1==a1, .. , ifelse( var1==a2, .. ,..))}
}
