% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{dim.matrix}
\alias{dim.matrix}
\title{return dim of symbolic Matrix}
\usage{
\method{dim}{matrix}(A)
}
\arguments{
\item{A}{Matrix}
}
\value{
dimension vector
}
\description{
return dim of symbolic Matrix
}
