% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.080morphism.domain.R
\name{dimension.of.domain}
\alias{dimension.of.domain}
\title{return dimension (number) of a domain expression}
\usage{
dimension.of.domain(domain)
}
\arguments{
\item{domain}{the domain expression}
}
\value{
the dimension of ambience space
}
\description{
return dimension (number) of a domain expression
}
