% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{dimension.of.image}
\alias{dimension.of.image}
\title{return the dimension of the image}
\usage{
dimension.of.image(m)
}
\arguments{
\item{m}{morphism}
}
\value{
dimension of the ambience space of the image
}
\description{
return the dimension of the image
}
