% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{distribution.common.new}
\alias{distribution.common.new}
\title{class for recording the distribution common distributions}
\usage{
distribution.common.new(family = "normal", dim = 1, mean = numeric(dim),
  cov. = diag(rep(1, dim)), ...)
}
\arguments{
\item{family}{distribution family}

\item{dim}{dimension}

\item{mean}{mean vector}

\item{cov.}{covariance matrix}

\item{...}{any thing you want to put to the detail}
}
\value{
a distribution
}
\description{
class for recording the distribution common distributions
}
