% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{disturbation.apply}
\alias{disturbation.apply}
\title{apply a disturbation on given fixed and random parameters}
\usage{
disturbation.apply(d, fixed, random)
}
\arguments{
\item{d}{disturbation}

\item{fixed}{fixed part}
}
\value{
expression
}
\description{
apply a disturbation on given fixed and random parameters
}
