% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05graphic.000common.R
\name{divide.graph}
\alias{divide.graph}
\title{divide graph into connected.components
this will go wrong if more vertex than edges
you have to put self loop if you want those vertex being component}
\usage{
divide.graph(G)
}
\arguments{
\item{G}{graph}
}
\value{
decomposition of vertex
}
\description{
divide graph into connected.components
this will go wrong if more vertex than edges
you have to put self loop if you want those vertex being component
}
