% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.080morphism.domain.R
\name{does.morphism.domain.contain}
\alias{does.morphism.domain.contain}
\title{check if a vector belongs to domain}
\usage{
does.morphism.domain.contain(m, vec)
}
\arguments{
\item{m}{morphism}

\item{vec}{vector}
}
\value{
TRUE or FALSE
}
\description{
check if a vector belongs to domain
}
