% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.002dynamic.system.R
\name{dynamic.system.new}
\alias{dynamic.system.new}
\title{description as kinetic.system}
\usage{
dynamic.system.new(ncompartment = 3, defn)
}
\arguments{
\item{ncompartment}{number of compartments}

\item{defn}{definition of dynamics, can be kinetic system or even general ODE}
}
\value{
skeleton
}
\description{
description as kinetic.system
}
