% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{edit.one.line.code.estimable.model}
\alias{edit.one.line.code.estimable.model}
\title{edit code}
\usage{
\method{edit}{one.line.code.estimable.model}(esm, lhs, rhs,
  which.part = "PK.PROGRAM")
}
\arguments{
\item{esm}{estimable.unit}

\item{lhs}{quoted, or character}

\item{rhs}{new definition}

\item{which.part}{which code, necessary}
}
\value{
new mode
}
\description{
pay attention, not all code are related to a parameter/estimable.unit/random.observation
some are just expression for computation
}
