% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\name{eliminate.constant.initilization.in.eqns}
\alias{eliminate.constant.initilization.in.eqns}
\title{remove constant initilization
if a initilization for constant is found, 
i.e. an equation which has no rhs symbols; 
or say, cannot be back referred by any edges
we just substitute the contant value of it to all appearence after it 
and remove the equation itself from equations}
\usage{
eliminate.constant.initilization.in.eqns(eqns)
}
\arguments{
\item{eqns}{equations}
}
\value{
equations whose contant initilization are removed
}
\description{
remove constant initilization
if a initilization for constant is found, 
i.e. an equation which has no rhs symbols; 
or say, cannot be back referred by any edges
we just substitute the contant value of it to all appearence after it 
and remove the equation itself from equations
}
