% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.007eliminate.rhs.para.R
\name{eliminate.duplicated.appearence.of.rhs.parameters}
\alias{eliminate.duplicated.appearence.of.rhs.parameters}
\title{eliminate duplicated appearence of NONMEM parameters}
\usage{
eliminate.duplicated.appearence.of.rhs.parameters(eqns)
}
\arguments{
\item{eqns}{list of equations}
}
\value{
equivalent equations
}
\description{
multiple appearence of THETA[i] or ETA[i] or EPS[i] as rhs in equations may cause problem
because of our assumption any appearence of THETA[] ETA[] EPS[] will create new estimable.unit 
or random observations
one simple solution is, if find multiple rhs THETA[1] then introduce a symbol parameter and replace all
appearence of THETA[1] by the introduced symbol
}
