% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\name{eliminate.loop.in.eqns}
\alias{eliminate.loop.in.eqns}
\title{add temporary variables to remove loops in equations}
\usage{
eliminate.loop.in.eqns(eqns)
}
\arguments{
\item{eqns}{equations}
}
\value{
equations which temporary variables are added to break the loops in generated graphs
}
\description{
add temporary variables to remove loops in equations
}
