% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\name{eliminate.na.initilization.in.eqns}
\alias{eliminate.na.initilization.in.eqns}
\title{eliminate constant initilization to NA
\code{ A = NA }
will be eliminated}
\usage{
eliminate.na.initilization.in.eqns(eqns)
}
\arguments{
\item{eqns}{equations}
}
\value{
eliminated equations
}
\description{
This will make code more beautiful, 
based on our assumption that nonmem will initialized user variable to NA
}
