% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\name{eliminate.one.equation.constant.initilization.in.eqns}
\alias{eliminate.one.equation.constant.initilization.in.eqns}
\title{used by eliminate.constant.initilization.in.eqns}
\usage{
eliminate.one.equation.constant.initilization.in.eqns(eqns, el.ind)
}
\arguments{
\item{eqns}{equations}

\item{el.ind}{the index (only one number is allowed) for which equation to be eliminated}
}
\value{
eliminated equations
}
\description{
used by eliminate.constant.initilization.in.eqns
}
