% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\name{eliminate.one.equation.variable.alias.in.eqns}
\alias{eliminate.one.equation.variable.alias.in.eqns}
\title{used by eliminate.variable.alias.in.eqns
both left hand side appearence or right hand side appearence of lhs of the alias will be 
substitute by rhs of alias 
hence, to ensure it's semanticly correct, we have to ensure the right hand side of \code{eqns[[el.ind]]}
is not used as right hand side in the following equations, 
or the value of the right hand side is distoried}
\usage{
eliminate.one.equation.variable.alias.in.eqns(eqns, el.ind)
}
\arguments{
\item{eqns}{equations}

\item{el.ind}{the index which equation should be removed}
}
\value{
eliminated equations
}
\description{
used by eliminate.variable.alias.in.eqns
both left hand side appearence or right hand side appearence of lhs of the alias will be 
substitute by rhs of alias 
hence, to ensure it's semanticly correct, we have to ensure the right hand side of \code{eqns[[el.ind]]}
is not used as right hand side in the following equations, 
or the value of the right hand side is distoried
}
