% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\name{eliminate.variable.alias.in.eqns}
\alias{eliminate.variable.alias.in.eqns}
\title{eliminate redundant variables
if theres is an alias format as following:}
\usage{
eliminate.variable.alias.in.eqns(eqns)
}
\arguments{
\item{eqns}{equations}
}
\value{
eliminated equations
}
\description{
\code{A = B}
\code{C = A}
then, eliminate A in the equations
we only deal with \code{symbol = symbol} (alias) pattern in equations
}
\details{
Of course, reusing some temporary varible will fool this function
but it will work well in simplifying the result of \code{compile.ifelse.to.if.statements}
}
