% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{embeded.ind.of.two.estimable.unit}
\alias{embeded.ind.of.two.estimable.unit}
\title{Get the index(position) of a parameter of a estimable unit when it's part of a model}
\usage{
embeded.ind.of.two.estimable.unit(esm, v1, v2, v1.subind = 1, v2.subind = 1)
}
\arguments{
\item{esm}{estimable model}

\item{v1}{estimable.unit 1}

\item{v2}{estimable.unit 2}

\item{v1.subind}{one estimable.unit may have more than one random varible in it, this indicate which variable are operated on}

\item{v2.subind}{similar to \code{v1.subind}}
}
\value{
embeded index and the corresponding backref (if available, or just NAs)
}
\description{
One estimable unit can be meaningful on it own.
But we usually see it as an embeded object in an estimable model.
When we list all the parameters (corresponding to those THETAs, ETAs) in a big table, it's useful to know their positions.
}
