% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.sample.R
\name{empty.sample}
\alias{empty.sample}
\title{Return an empty data.frame with the columns returned by \code{random.search}}
\usage{
empty.sample()
}
\value{
An empty data.frame
}
\description{
Return an empty data.frame with the columns returned by \code{random.search}
}
\seealso{
random.search
}
