% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{equal.symbol.light}
\alias{equal.symbol.light}
\title{equal.symbol.light}
\usage{
equal.symbol.light(e1, e2)
}
\arguments{
\item{e1}{expression}

\item{e2}{expression}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
light weighted testing of equality \cr
Can ensure if the return value is \code{TRUE}, \code{e1} must equal to \code{e2};  \cr
however, it's not able to ensure when \code{e1} equals to \code{e2} mathematically, the return value is \code{TRUE}
}
\author{
jjxie
}
