% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{estimable.Unit.correct.lognormal}
\alias{estimable.Unit.correct.lognormal}
\title{correct log normal case when the value is defined by \code{x -> log(x)}}
\usage{
estimable.Unit.correct.lognormal(e)
}
\arguments{
\item{e}{estimable unit}
}
\value{
fixed version
}
\description{
correct log normal case when the value is defined by \code{x -> log(x)}
}
