% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{estimable.Unit.new}
\alias{estimable.Unit.new}
\title{Promotion a expression to estimable unit}
\usage{
estimable.Unit.new(m)
}
\arguments{
\item{m}{morphism as estimable unit}
}
\value{
m with class attribute updated
}
\description{
promote a morphism or a expression to a estimable.unit
However, 'name' is special in R, if an class attribute is added to a name, R will be forced to evaluate it,
need to avoid this.
}
