% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.003estimable.model.R
\name{eval.backref.from.var}
\alias{eval.backref.from.var}
\title{extract the INDEX of the \code{ETA} entries}
\usage{
eval.backref.from.var(e)
}
\arguments{
\item{e}{variance expression}
}
\value{
the index
}
\description{
internal use, Only on \code{ETA}
}
