% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.050rnmimport.R
\name{eval.nonmem.parameter}
\alias{eval.nonmem.parameter}
\title{eval.nonmem.parameter}
\usage{
eval.nonmem.parameter(e0, md)
}
\arguments{
\item{e0}{expression}

\item{md}{NONMEM model generated by \code{importNmMod}}
}
\value{
evaluated expression
}
\description{
evaluate expression including NONMEM parameters\cr
assuming \code{md} is a RNMImport model returned by \code{importNmMod} \cr
\code{eval.nonmem.parameter} can evaluate an expression w.r.t the model's parameter's definition\cr
e.g.\cr
    \code{eval.nonmem.parameter( quote( Var(ETA[1]) ) , md)          -> Omega[1,1] } \cr
    \code{eval.nonmem.parameter( quote( Cov(ETA[1], ETA[2]) ) , md)  -> Omega[1,2] } \cr
}
\author{
Mango solutions
}
