% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{eval.symbolic}
\alias{eval.symbolic}
\title{eval.symbolic}
\usage{
eval.symbolic(e, eval.env = new.env(parent = emptyenv()))
}
\arguments{
\item{e}{expression or list of expression}

\item{eval.env}{evaluation environment}
}
\value{
last value of evaluation
}
\description{
More handy interface for \code{\link{eval.symbolic0}} \cr
input can now be a expression or several code in \code{{}}'s
}
\author{
jjxie
}
