% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{eval.symbolic0}
\alias{eval.symbolic0}
\title{eval.symbolic0
evaluate a sequence of equations}
\usage{
eval.symbolic0(eqns, eval.env = new.env(parent = emptyenv()))
}
\arguments{
\item{eqns}{equations (list of expressions)}

\item{eval.env}{the environment for evaluation}
}
\value{
the last value of evaluation
}
\description{
\code{ TVKA = THETA[1]*WT/70 }\cr
\code{ KA = TVKA * exp(ETA[1]) } \cr
}
\details{
can be evaluated to \code{ KA= THETA[1]*WT/70 * exp(ETA[1]) }
}
\author{
jjxie
}
