% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exaustive.search.R
\name{exaustive.search}
\alias{exaustive.search}
\title{Explore all combinations of formula terms}
\usage{
exaustive.search(
  complete.X.df,
  y,
  n.squares = 1,
  formula.len = 3,
  K = 7,
  N = 10,
  seed = NULL,
  transformations = list(log10 = function(rdf, x, z) {
     log10(0.1 + abs(z) + x)
 },
    inv = function(rdf, x, z) {
     1/(0.1 + abs(z) + x)
 }),
  custom.abs.mins = list(),
  glob.filepath = NULL,
  chunk.size = NULL,
  cv.norm = FALSE
)
}
\arguments{
\item{complete.X.df}{The dataset that contains the base variables the formula is composed of (column-wise)}

\item{y}{The independent variable to be predicted with the formula}

\item{n.squares}{The maximum order of the polynomial composition of base variables. Ex. \verb{order 0 = a}, \verb{order 1 = a*b}, \verb{order 2 = a*b*c}}

\item{formula.len}{The number of terms in the formulas that will be randomly sampled.}

\item{K}{The number of parts the dataset is split into for K-fold cross-validation.}

\item{N}{The number of times the K-fold validation is repeated, shuffling the dataset row orders before each time.}

\item{seed}{An (optional) seed for deterministic run}

\item{transformations}{A list of potentially non-linear transformations that can be applied on top of the squares. Ex. \verb{order 0, transformation=log10 = log10.a}}

\item{custom.abs.mins}{A list of user-defined minimum values for dataset columns.}

\item{glob.filepath}{The path to an rDdata object containing the results of potentially multiple independent previous run, that will be excluded from the current run. The file will be automatically updated whith the new formula evaluation results.}

\item{chunk.size}{If null, compute all missing formulas. If a number, compute only that number of missing formulas.}

\item{cv.norm}{Normalize regressors after train-validation split in inner cross-validation loop.}
}
\value{
A data.frame of formulas and the corresponding cross-validation performance measures (R-squared, absolute relative error, max cooks distance). See also \code{empty.sample}.
}
\description{
Explore all combinations of formula terms
}
\examples{
\donttest{
# set-up a toy example dataset
x1<-runif(100, min=2, max=67)
x2<-runif(100, min=0.01, max=0.1)
X <- data.frame(x1=x1, x2=x2)
# set up a "true" non-linear relationship
# with some noise
y <- log10(x1^2*x2) + rnorm(100, 0, 0.001)

max.formula.len=formula.len=1
n.squares=1
K=2
N=2
seed=1001
res.new <- exaustive.search(X, y,
                            n.squares=1,
                            formula.len=3,
                            K=2, N=3, seed=NULL,
                            transformations=list(
               "log"=function(rdf, x, stats){ log(x) },
               "log_x1_p"=function(rdf, x, stats){ log(rdf$x1 + x) },
               "inv"=function(rdf, x, stats){ 1/x }
                            ),
                            custom.abs.mins=list(),
                            glob.filepath = NULL,
                            chunk.size=NULL, cv.norm=FALSE)
}
}
