% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{expand.as.product}
\alias{expand.as.product}
\title{expand.as.product}
\usage{
expand.as.product(e)
}
\arguments{
\item{e}{expression}
}
\value{
list of expression
}
\description{
Expand the production into list, \code{A * B * C * D} \code{->} \code{[ A, B, C ,D]} \cr
also some extra translation is applied:   \cr
                                         \code{-A -> [-1,A] } \cr
                                         \code{A/B -> [A, B^-1] } \cr
                                         \code{A^1 -> [A] , (A^i)^j -> [A^i*j]}
}
\author{
jjxie
}
