% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{expand.as.sum.of.product}
\alias{expand.as.sum.of.product}
\title{expand.as.sum.of.product}
\usage{
expand.as.sum.of.product(e)
}
\arguments{
\item{e}{expression}
}
\value{
list of expressions
}
\description{
expand an expression into list of list as following \cr
  \code{ A*B + C*D + E*F -> [ [ A, B] , [C, D] , [E, F] ] }
}
\details{
non-deep-expand
 can do this : \code{(a+b)*c -> [ [a+b, c] ]}
 cannot do this : \code{(a+b)*c -> a*c + b*c}
}
\seealso{
\code{\link{print.listoflist}}
}
\author{
jjxie
}
