% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{expand.as.sum}
\alias{expand.as.sum}
\title{expand.as.sum}
\usage{
expand.as.sum(e)
}
\arguments{
\item{e}{expression}
}
\value{
list of expression
}
\description{
Expand the summation into list, \cr 
\code{A + B + C + D} \samp{->} \code{[ A, B, C ,D]}
}
\author{
jjxie
}
