% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\name{expand.block.to.list}
\alias{expand.block.to.list}
\title{a expression itselt is changed into \code{list(e)}}
\usage{
expand.block.to.list(e)
}
\arguments{
\item{e}{expression}
}
\value{
a list of equation , any block '{ ... }' is expaned into list
}
\description{
a expression itselt is changed into \code{list(e)}
}
