% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.110morphism.combinators.R
\name{explain.expression.as.morphism}
\alias{explain.expression.as.morphism}
\title{THETA oriented function
This try to guess a meaning of an expression contain THETAs
we presume the THETA's as the points in the preimage
however, we discard the informations here, i.e. remove THETAs here}
\usage{
explain.expression.as.morphism(e)
}
\arguments{
\item{e}{expression}
}
\value{
a morphism object
}
\description{
THETA oriented function
This try to guess a meaning of an expression contain THETAs
we presume the THETA's as the points in the preimage
however, we discard the informations here, i.e. remove THETAs here
}
