% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.110morphism.combinators.R
\name{explain.morphism.as.morphism}
\alias{explain.morphism.as.morphism}
\title{just an application of \code{explain.expression.as.morphism}
instantiate \code{m} using THETAs
then explain it
might be wrong if THETAs have other meaning in \code{m} than the arguments}
\usage{
explain.morphism.as.morphism(m)
}
\arguments{
\item{m}{morphism}
}
\value{
morphisms
}
\description{
just an application of \code{explain.expression.as.morphism}
instantiate \code{m} using THETAs
then explain it
might be wrong if THETAs have other meaning in \code{m} than the arguments
}
