% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{exponential.matrix.of.monomials.sum.product}
\alias{exponential.matrix.of.monomials.sum.product}
\title{exponential.matrix.of.monomials.sum.product}
\usage{
exponential.matrix.of.monomials.sum.product(le)
}
\arguments{
\item{le}{list of list of expression}
}
\value{
matrix of exponentials
}
\description{
For a list of list as following: \cr
\code{ le : [ [ A^2 ,B, A] ,...] } \cr
generate the following matrix: \cr
\verb{
   .   A  B  ...  
   1   3  1  0   
   2   ......... 
}\cr
each row for a entry of list \code{le} \cr
each column is a monomial appearing in the whole list\cr
each entry of matrix is the degree of the monomial
}
\author{
jjxie
}
