% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50pattern.002error.default.R
\name{export.error.default}
\alias{export.error.default}
\title{export.error.default}
\usage{
export.error.default(errorObjs)
}
\arguments{
\item{errorObjs}{pattern found by \code{pattern.error.default}}
}
\value{
list of text of \code{$ERROR} block and mappings for \code{EPS}
}
\description{
Export pattern found by \code{\link{pattern.error.default}}
}
\author{
Mango solutions
}
