% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50pattern.001pk.default.R
\name{export.pk.default}
\alias{export.pk.default}
\title{export.pk.default}
\usage{
export.pk.default(pkObjs)
}
\arguments{
\item{pkObjs}{pattern recognized by \code{pattern.pk.default}}
}
\value{
list containing the text for \code{$PK} block and the mappings for \code{THETA} and \code{ETA}
}
\description{
Exporter for the above pattern found by \code{\link{pattern.pk.default}}
}
\author{
Mango solutions
}
