% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{expression.from.list.of.sum}
\alias{expression.from.list.of.sum}
\title{expression.from.list.of.sum}
\usage{
expression.from.list.of.sum(le)
}
\arguments{
\item{le}{list of expression}
}
\value{
one expression
}
\description{
Convert back list of sum to R expression\cr
inverse of \code{\link{expand.as.sum}}, from a list of expression, create a single expression
}
\author{
jjxie
}
