% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{expression.from.sum.product}
\alias{expression.from.sum.product}
\title{expression.from.sum.product}
\usage{
expression.from.sum.product(le)
}
\arguments{
\item{le}{list of expressions}
}
\value{
one expression
}
\description{
Convert back list of sum of list of product to R expression \cr
e.g. \code{[ [a1,a2,a3] , [b1,b2,b3], [c1] ] -> a1*a2*a3 + b1*b2*b3 + c1}
}
\seealso{
\code{\link{expand.as.sum.of.product}}
}
\author{
jjxie
}
