% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.000common.R
\name{expressions.has.random.variables}
\alias{expressions.has.random.variables}
\title{test if the expression has some \code{ETA[i]} or \code{EPS[i]}}
\usage{
expressions.has.random.variables(e)
}
\arguments{
\item{e}{expression}
}
\value{
logical value
}
\description{
test if the expression has some \code{ETA[i]} or \code{EPS[i]}
}
