% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.070constructor.R
\name{extract.coefficient.of.variance}
\alias{extract.coefficient.of.variance}
\title{extract.coefficient.of.variance}
\usage{
extract.coefficient.of.variance(Var)
}
\description{
assuming the Var part are of pattern coeff * Var(ETA[i]) or just Var(ETA[i])
 for expression \code{coeff * Var(ETA[i])} or \code{coeff * Var(EPS[i])}
 for generating: \code{ rnd.part = sqrt(coeff) * ETA[i] }, we need the \code{coeff}
}
\details{
@param Var variance representation
 @return found coefficient for standard deviation
}
