% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{extract.edges.quotient.array}
\alias{extract.edges.quotient.array}
\title{extract.edges.quotient.array}
\usage{
extract.edges.quotient.array(e0)
}
\arguments{
\item{e0}{parsed R expression}
}
\value{
list of lhs and rhs
}
\description{
From one equation \code{LHS = RHS}
extract several symbols in the right hand side \cr
and symbol in left hand side \cr
the \code{array[i]} \code{array[j]} will be extract just as \code{array[]} \cr
i.e. the set of edges are quotient by the equivlent relation of glueing together all \cr
array entries into a single array name, hence the vertex.table are only characters
}
\author{
jjxie
}
