% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.003estimable.model.R
\name{extract.random.index}
\alias{extract.random.index}
\title{extract index of ETA or EPS}
\usage{
extract.random.index(e)
}
\arguments{
\item{e}{expression}
}
\value{
index
}
\description{
extract index of ETA or EPS
}
