% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{extract.vertices}
\alias{extract.vertices}
\title{extract.vertices}
\usage{
extract.vertices(e0)
}
\arguments{
\item{e0}{parsed R expression}
}
\value{
list of lhs and rhs, which are just expressions
}
\description{
From one equation \code{LHS = RHS}
extract several symbols in the right hand side \cr
and symbol in left hand side \cr
the array[i] array[j] will be extract just as array[]
}
\details{
Those symbols lhs and rhs are actually vertices in a relation graph.
If the expression is a simple equal relation, the graph are the complete bipartite graph, i.e.
   all pairs \code{ <rhs[i],lhs[j]> }, for any \code{i,j}.

If the expression is a "if" statement, and the "yes" "no" part contain multiple statements,
   the edges are not the all pairs as above. Should use \code{create.abstractGraph.from.equations}
   to get the actual edges.
}
\author{
jjxie
}
