% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{extractBalanced}
\alias{extractBalanced}
\title{extractBalanced}
\usage{
extractBalanced(stream, leadingChar = "IF\\\\s*")
}
\arguments{
\item{stream}{a string}

\item{leadingChar}{character pattern}
}
\value{
a character with length 2, first is the extract string, second is the left stream
}
\description{
extract balanced Tokens : () '' "" ,e.t.c just after the \code{leadingChar} \cr
used by \code{\link{parse.if}}
}
\details{
When there is any nested \code{()}'s, should it be carefully counted.
}
\author{
Mango solutions
}
