% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{factorize.2}
\alias{factorize.2}
\title{factorize.2}
\usage{
factorize.2(e)
}
\arguments{
\item{e}{expression}
}
\value{
an expression as production
}
\description{
very light weighted factorization \cr
just factor out commond terms \cr
however, we haven't used any division algorithm's here, \cr
hence even can't recognized \code{A|A^k}
}
\author{
jjxie
}
