% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{factorize.3}
\alias{factorize.3}
\title{factorize.3}
\usage{
factorize.3(e)
}
\arguments{
\item{e}{expression}
}
\value{
an expression as production
}
\description{
better than factorize.2,
can recognize A A^k has common divisors
}
\author{
jjxie
}
