% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60knowledge.001subroutine.R
\name{find.socp.symbols.by.subroutine}
\alias{find.socp.symbols.by.subroutine}
\title{from equations, find possible PK parameters
The reason why we need \code{eqns} here is because alias
we actually do not want all possible ones}
\usage{
find.socp.symbols.by.subroutine(eqns, subs = "ADVAN4", trans = "",
  try_abbr_pk_names = TRUE)
}
\arguments{
\item{eqns}{equations}

\item{subs}{PREDPP subroutine name}

\item{trans}{translation name used}

\item{try_abbr_pk_names}{should try match ALPH to ALPHA, e.t.c}
}
\value{
list of found PK parameters
}
\description{
from equations, find possible PK parameters
The reason why we need \code{eqns} here is because alias
we actually do not want all possible ones
}
