% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{fortran2r}
\alias{fortran2r}
\title{fortran2r}
\usage{
fortran2r(x)
}
\arguments{
\item{x}{string}
}
\value{
string
}
\description{
translate NONMEM operators to R valid operators\cr
this function won't check if the pattern is in the quoted case, if your string has \cr
\code{"as .EQ. "} , it will be destoried as \code{"as == "}
}
\author{
Mango solutions
}
