% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{generate.formal.arguments.morphism}
\alias{generate.formal.arguments.morphism}
\title{generate formal arguments, which not existing in the symbols}
\usage{
generate.formal.arguments.morphism(m, fsymbol = quote(x))
}
\arguments{
\item{m}{morphism}

\item{fsymbol}{symbol used for arguments, e.g. THETA or x}
}
\value{
list of indeterminators
}
\description{
generate formal arguments, which not existing in the symbols
}
