% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{generate.func.firstorder.linear.ode}
\alias{generate.func.firstorder.linear.ode}
\title{generate deSolve usable derivatives function}
\usage{
generate.func.firstorder.linear.ode(ode)
}
\description{
generate deSolve usable derivatives function
}
