% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{generate.unused.symbol.in.expressions}
\alias{generate.unused.symbol.in.expressions}
\title{for one or more expressions
find in \code{available.symbol.table} an not used one, which can represent the formal arguments}
\usage{
generate.unused.symbol.in.expressions(es, available.symbol.table = c("x", "y",
  "z", "w", "a", "alpha", "beta", "gamma"))
}
\arguments{
\item{es}{one or more expressions}

\item{available.symbol.table}{characters which looks better for printing}
}
\value{
first found symbol, or NULL if not found
}
\description{
for one or more expressions
find in \code{available.symbol.table} an not used one, which can represent the formal arguments
}
