% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic.search.R
\name{genetic.search}
\alias{genetic.search}
\title{Genetic Algorithm for non-linear formula optimization}
\usage{
genetic.search(
  complete.X.df,
  y,
  n.squares = 1,
  max.formula.len = 4,
  seed = NULL,
  fitness.fun = pe.r.squared.formula.len.fitness,
  transformations = list(log10 = function(rdf, x, z) {
     log10(0.1 + abs(z$min) + x)

    }, inv = function(rdf, x, z) {
     1/(0.1 + abs(z$min) + x)
 }),
  custom.abs.mins = list(),
  glob.filepath = NULL,
  local.filepath = NULL,
  memoization = F,
  monitor = monitor.formula.fun,
  maxiter = 100,
  N = 2,
  K = 7,
  pcrossover = 0.2,
  popSize = 50,
  pmutation = 0.8,
  keepBest = F,
  cv.norm = F,
  best.vars.l = list()
)
}
\arguments{
\item{complete.X.df}{The dataset that contains the base variables the formula is composed of (column-wise)}

\item{y}{The independent variable to be predicted with the formula}

\item{n.squares}{The maximum order of the polynomial composition of base variables. Ex. \verb{order 0 = a}, \verb{order 1 = a*b}, \verb{order 2 = a*b*c}}

\item{max.formula.len}{The maximum number of terms in the formula}

\item{seed}{An (optional) seed for deterministic run}

\item{fitness.fun}{The function that determine the fitness of a given formula. Defaults to \code{pe.r.squared.formula.len.fitness}}

\item{transformations}{A list of potentially non-linear transformations that can be applied on top of the squares. Ex. \verb{order 0, transformation=log10 = log10.a}. Input values are x and z, the array of numbers to be transformed (training set only), and the min max statistics (on the global dataset) respectively.}

\item{custom.abs.mins}{A list of user-defined minimum values for dataset columns.}

\item{glob.filepath}{Has effect only if memoization=TRUE. The path to an rDdata object containing the results of potentially multiple independent previous run.}

\item{local.filepath}{Has effect only if memoization=TRUE. The path to an rData object where the results of the current run will be stored. If it already exists, the new results will be appended.}

\item{memoization}{If TRUE test results will be stored in \code{res.filepath}}

\item{monitor}{Function that will be called on every iteration with the current solutions. Defaults to \code{monitor.formula.fun}}

\item{maxiter}{Maximum number of genetic evolution epochs}

\item{N}{The number of times the K-fold validation is repeated, shuffling the dataset row orders before each time.}

\item{K}{The number of parts the dataset is split into for K-fold cross-validation.}

\item{pcrossover}{The probability of crossover between pairs of chromosomes. Typically this is a large value and by default is set to 0.8.}

\item{popSize}{The population size, the number of formulas considered for genetic evolution.}

\item{pmutation}{The probability of mutation in a parent chromosome. Usually mutation occurs with a small probability, and by default is set to 0.1.}

\item{keepBest}{If TRUE, the value \code{best.iter} of the list returned, will contain a list of the best formulas at each evolution iteration.}

\item{cv.norm}{Normalize regressors after train-validation split in inner cross-validation loop.}

\item{best.vars.l}{A list of formulas. Each formula is an array of strings, each string is the textual representation of a formula term. Ex. \code{cur.vars.l <- list(c('a','mul.a.b'))} will test the formula \code{y ~ a + a*b}. This list is used as a starting point for genetic evolution. It may come from a-priori knowledge or by extracting the most promising results from \code{random.search}.}
}
\value{
A list with three values:
list(
best: the best formula found overall
best.iter: a list of best formulas, one for each evolution iteration
results: The output of the GA::ga function, with all the solutions, hyperparameters, etc.
)
}
\description{
Starting from an (optional) list of promising formulas,
generated with other methods such as \code{random.search},
explore combinations (crossover) and alterations (mutation)
of them in order maximize the value of the fitness function
that defaults to \code{pe.r.squared.formula.len.fitness}
}
\examples{
\donttest{
# set-up a toy example dataset
x1<-runif(100, min=2, max=67)
x2<-runif(100, min=0.01, max=0.1)
X <- data.frame(x1=x1, x2=x2)
# set up a "true" non-linear relationship
# with some noise
y <- log10(x1^2*x2) + rnorm(100, 0, 0.001)

# stats:
#    list(
#       "min"=..,
#       "absmin"=..,
#       # zero in original space projected in std space
#       "projzero"=prjzero
#    )
transformations=list(
  "log"=function(rdf, x, stats){ log(x) },
  "log_x1_p"=function(rdf, x, stats){ log(rdf$x1 + x) },
  "inv"=function(rdf, x, stats){ 1/x }
)

# empty datastore
datastore='test.rData'
genetic.datastore='test.genetic.rData'
random.datastore='test.random.rData'
saveRDS(empty.sample(), datastore)
saveRDS(empty.sample(), genetic.datastore)
saveRDS(empty.sample(), random.datastore)

random.res <- random.search(
     X, y,
     n.squares=2,
     formula.len=1,
     maxiter=10,
     # formulas from previous runs
     glob.filepath = datastore,
     # formulas computed by this call
     local.filepath = random.datastore,
     transformations = transformations,
     memoization=TRUE
 )

res <- readRDS(datastore)
# check random.res...
# append to global results to avoid recompute
res <- rbind(res, random.res)
saveRDS(res, datastore)

# search more with genetic algorithm without
# re-analyzing already analyzed formulas
# stored in datastore
best.finetuned <- genetic.search(
 X,
 y,
 n.squares=1,
 maxiter=10,
 glob.filepath=datastore,
 local.filepath=genetic.datastore,
 memoization=TRUE,
 pcrossover=0.2,
 pmutation=0.8,
 seed=NULL,
 max.formula.len=4,
 keepBest=TRUE,
 K=2,
 N=3,
 popSize = 5,
 fitness.fun=pe.r.squared.formula.len.fitness,
 # multiple best initial guess
 best.vars.l = list(
       c('mul.x1.x2','x2')
 )
)

file.remove(datastore)
file.remove(genetic.datastore)
file.remove(random.datastore)
}
}
\seealso{
random.search
}
