% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{get.basic.pk.parameters}
\alias{get.basic.pk.parameters}
\title{For dynamics description, return the necessary basic pk parameters}
\usage{
get.basic.pk.parameters(dyn)
}
\arguments{
\item{dyn}{dynamics}
}
\value{
basic pk parameters
}
\description{
The basic pk parameters are necessary for NONMEM PREDPP subroutines to run.
Any lack of those parameters will make NONMEM complain.
}
