% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.003estimable.model.R
\name{get.by.ind.triple.matrix}
\alias{get.by.ind.triple.matrix}
\title{Ensure return value of subsetting from a \code{triple.matrix} to be non-\code{NA}}
\usage{
get.by.ind.triple.matrix(ind0, triple.matrix, columns = 1:3,
  default.val.for.col = c(-Inf, 1e-04, Inf), remove.original.nas = TRUE)
}
\arguments{
\item{ind0}{original indices}

\item{triple.matrix}{typically a Theta in nmmod}

\item{columns}{select which columns}

\item{default.val.for.col}{default value for specified column for non-existing indices}

\item{remove.original.nas}{if there are NAs in the input matrix, we set those to default values}
}
\value{
selected array
}
\description{
If \code{ind0}, the index usr provide is \code{NA}, the return is just the \code{default.val.for.col} default values. 
e.g. For column one (lower bound) is minus infinity and upper bound positive infinity and the initial estimation to be a small positive number(NOT zero! NONMEM will complain)
}
\details{
If \code{triple.matrix} itself has \code{NA}s, and \code{remove.original.nas} is \code{TRUE}, 
those \code{NA}s are also set to \code{default.val.for.col} by their column position.
}
