% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{get.cov.estimable.model}
\alias{get.cov.estimable.model}
\title{return covariance matrix of all random parameters for the model spec}
\usage{
get.cov.estimable.model(esm, indexed.by = "subject")
}
\arguments{
\item{esm}{estimable.model}

\item{indexed.by}{subject or observation random effect}
}
\value{
named covariance matrix
}
\description{
return covariance matrix of all random parameters for the model spec
}
