% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.003eqns.R
\name{get.definition.of.v}
\alias{get.definition.of.v}
\title{get.definition.of.v}
\usage{
get.definition.of.v(eqns, vn)
}
\arguments{
\item{eqns}{equations}

\item{vn}{symbol of variable}
}
\value{
evaluated value of \code{vn} using \code{eqns} and the depending variables
}
\description{
find necessary equations for a given variable
}
