% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{get.domain.estimable.unit}
\alias{get.domain.estimable.unit}
\title{extract the domain part of a estimable.unit}
\usage{
get.domain.estimable.unit(m)
}
\value{
domain expression
}
\description{
Only fixed part has domain.
}
