% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{get.initial.estimation.estimable.unit}
\alias{get.initial.estimation.estimable.unit}
\title{get the initial estimation from a estimable unit}
\usage{
get.initial.estimation.estimable.unit(m)
}
\arguments{
\item{m}{estimable unit}
}
\value{
value of initial estimation
}
\description{
get the initial estimation from a estimable unit
}
