% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30pattern.001analyseConnected.R
\name{get.parameter.space.embedding.for.cp}
\alias{get.parameter.space.embedding.for.cp}
\title{get.parameter.space.embedding.for.cp}
\usage{
get.parameter.space.embedding.for.cp(cp)
}
\arguments{
\item{cp}{one connected component}
}
\value{
a list with modified cp(you may use or not use it) and the theta.hat and eta.hat defined by theta and eta
}
\description{
The meanning for this function is the following:
}
\details{
For a connected component, e.g.

\code{ V = A + ETA[1] }
\code{ CL = B + ETA[1] }

convert into

\code{ V = A + ETA.HAT[1] }
\code{ CL = B + ETA.HAT[2] }

where an extra embedding from one dimentional space \code{ETA[1]} to the two dimensional space \code{(ETA.HAT[1], ETA.HAT[2])}
with the mapping function

\code{ ETA[1] --> (ETA[1], ETA[1]) }
}
