% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.002dynamic.system.R
\name{get.subs.trans.dynamics}
\alias{get.subs.trans.dynamics}
\title{convert description in kinetic to NONMEM PREDPP subroutine, as characters}
\usage{
get.subs.trans.dynamics(d)
}
\arguments{
\item{d}{kinetic description}
}
\value{
strings of subs and trans
}
\description{
convert description in kinetic to NONMEM PREDPP subroutine, as characters
}
