% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{get.triple.estimable.unit}
\alias{get.triple.estimable.unit}
\title{the lower est upper triple}
\usage{
get.triple.estimable.unit(m)
}
\arguments{
\item{m}{estimable.unit}
}
\value{
triple
}
\description{
Usually we use a triple (three column data.frame) for storing the lower bound, estimation value and upper bound.
After extracting the comain of estimable unit, we convert it to the triple representation.
}
