% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympy_import.R
\name{get_sympy}
\alias{get_sympy}
\title{Get the sympy module cached in a R session}
\usage{
get_sympy(refresh = FALSE)
}
\arguments{
\item{refresh}{Logical value to indicate whether to force importing and refreshing the cache.}
}
\value{
The function returns the sympy module.
}
\description{
The sympy module is cached to `options("symbolicR.sympymoudle")`, this function
will check whether the option exists, if not, it will use `reticulate::import`
to import sympy and cache it.
}
