% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{guess.ode.time.variable}
\alias{guess.ode.time.variable}
\title{guess time variable}
\usage{
guess.ode.time.variable(vc)
}
\arguments{
\item{vc}{\code{state.vector} functions list \code{ [ x1(t), x2(t) ] }}
}
\value{
the names of functions and the time variable
}
\description{
guess time variable
}
