% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.000common.R
\name{guess.variable.meaning.by.name.default}
\alias{guess.variable.meaning.by.name.default}
\title{guess.variable.meaning.by.name.default}
\usage{
guess.variable.meaning.by.name.default(v)
}
\arguments{
\item{v}{variable name}
}
\value{
type
}
\description{
for a list or character, output the corresponding possible type of the variable
}
