% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.000common.R
\name{guess.variable.type.by.graph}
\alias{guess.variable.type.by.graph}
\title{these types doesn't have pk meaning\cr
even tempvar can be some PK parameter, e.g.\cr
if we added a line\cr}
\usage{
guess.variable.type.by.graph(G, v = "V")
}
\arguments{
\item{G}{graph of equations}

\item{v}{variable name}
}
\value{
type of variable in the graph
}
\description{
\code{S2 = V}
}
\details{
then, \code{V} becomes a tempvar, not necessary a terminal
}
