% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\docType{data}
\name{ifelse.statement.to.function.rules}
\alias{ifelse.statement.to.function.rules}
\title{convert if then else statements to function}
\format{An object of class \code{list} of length 3.}
\usage{
ifelse.statement.to.function.rules
}
\description{
But we create some loop here as
Q = if (cond) A else Q 
i.e. Q -> Q
}
\details{
hence we need to remove those loops
}
\keyword{datasets}
