% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{init_syms}
\alias{init_syms}
\title{Initiate Symbols}
\usage{
init_syms(..., .envir = parent.frame(), .quite = FALSE)
}
\arguments{
\item{...}{For arguments with a name, they will be treated as assumptions and pass them
to sympy, for arguments without a name, they are substituted as symbol names
to initiate. See example.}

\item{.envir}{The environment to initiate the symbols.}

\item{.quite}{Logical value to indicate whether to show messages when initiating the symbols}
}
\value{
It invisibly returns a list of the initiated symbols.
}
\description{
This is a convenience function to initiate symbols.
}
\examples{
init_syms(x, y, z, positive = TRUE, .quite = TRUE)
x
class(x)
GreaterThan(x, 0)
}
