% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25plainmath.001matrix.R
\name{inject.names.into.list.of.matrix}
\alias{inject.names.into.list.of.matrix}
\title{assign matrix names}
\usage{
inject.names.into.list.of.matrix(ml, dim.names)
}
\arguments{
\item{ml}{matrix list}

\item{dim.names}{overall names}
}
\value{
matrix list
}
\description{
assign matrix names
}
