% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{insert.one.line.code.estimable.model}
\alias{insert.one.line.code.estimable.model}
\title{insert code into model spec}
\usage{
insert.one.line.code.estimable.model(esm, lhs, rhs, which.part = "PK.PROGRAM",
  where.ind = 1)
}
\arguments{
\item{esm}{estimable.model}

\item{lhs}{the lhs of inserted code, should be quoted expression}

\item{rhs}{definition of lhs, can be a expression or estimable.unit}

\item{which.part}{insert to which part, PK or Observation PROGRAM}

\item{where.ind}{index, can be numeric, or character}
}
\value{
updated model
}
\description{
Handy character representation of place.
\code{^} refer to the beginning of code, \code{$} means appending after the code
\code{^CL} or \code{CL} means insert the new code before \code{CL} in the existing code
\code{CL$} means after that code
}
