% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{insert.one.line.code.list}
\alias{insert.one.line.code.list}
\title{insert code into a given list}
\usage{
insert.one.line.code.list(le, lhs, rhs, e, index)
}
\arguments{
\item{le}{list of code}

\item{lhs}{lhs symbol}

\item{rhs}{rhs definition, can be estimable unit}

\item{e}{assignment expression}

\item{index}{where to insert, should be number}
}
\value{
new list
}
\description{
Insert one line code into list. The code is an assignment, which the \code{lhs} and \code{rhs} should be given explicitly.
The \code{rhs} can be an atomic or a symbol or an expression or supported object like \code{estimable.unit} and \code{random.observation}.
}
\details{
If both \code{lhs} and \code{rhs} are missing, you have to provide an assignment expression in \code{e} and the function
can split it into lhs and rhs.
}
