% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{instantiate.morphism}
\alias{instantiate.morphism}
\title{for a given morphism export as one or more equations
just for applying on parameters
but not for currying
You can instantiate using THETAs as parameters, and then \code{explain.expression.as.morphism} to analyse
its structure}
\usage{
instantiate.morphism(m, parameter.pool = NULL)
}
\arguments{
\item{m}{morphism}

\item{parameter.pool}{e.g. \code{THETA[i]} which will be used for constructing the output mapping}
}
\value{
equations
}
\description{
for a given morphism export as one or more equations
just for applying on parameters
but not for currying
You can instantiate using THETAs as parameters, and then \code{explain.expression.as.morphism} to analyse
its structure
}
