% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.002dynamic.system.R
\name{kinetic.system.predpp.new}
\alias{kinetic.system.predpp.new}
\title{create a predpp known kinetic system}
\usage{
kinetic.system.predpp.new(s)
}
\arguments{
\item{s}{description from NONMEM model}
}
\value{
kinetic system
}
\description{
create a predpp known kinetic system
}
