% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{list.all.fixed.parameters.estimable.model}
\alias{list.all.fixed.parameters.estimable.model}
\title{list all fixed parameters in a model spec}
\usage{
list.all.fixed.parameters.estimable.model(esm, with.triple = FALSE)
}
\arguments{
\item{esm}{estimable.model}

\item{with.triple}{output lower est upper of fixed parameters}
}
\value{
data.frame
}
\description{
list all fixed parameters in a model spec
}
