% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{list.all.fixed.parameters.list}
\alias{list.all.fixed.parameters.list}
\title{list all parameters of fixed part in all \code{estimable.unit} from a list}
\usage{
list.all.fixed.parameters.list(le)
}
\arguments{
\item{le}{list of code}
}
\value{
data.frame
}
\description{
list all parameters of fixed part in all \code{estimable.unit} from a list
}
