% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{list.all.parameters.estimable.model}
\alias{list.all.parameters.estimable.model}
\title{list all paramters in one model spec}
\usage{
list.all.parameters.estimable.model(esm)
}
\arguments{
\item{esm}{estimable.model}
}
\value{
data.frame
}
\description{
list all paramters in one model spec
}
