% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.000common.R
\name{list.any.is.nonmem.par}
\alias{list.any.is.nonmem.par}
\title{test any entry of the list is a nonmem parameters
if any element of \code{le} is a nonmem parameter \code{THETA}, \code{ETA} or \code{EPS},
return \code{TRUE}, else \code{FALSE}}
\usage{
list.any.is.nonmem.par(le)
}
\arguments{
\item{le}{list of expressions}
}
\value{
logical value
}
\description{
test any entry of the list is a nonmem parameters
if any element of \code{le} is a nonmem parameter \code{THETA}, \code{ETA} or \code{EPS},
return \code{TRUE}, else \code{FALSE}
}
