% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{lookup.random.part.backref.estimable.unit}
\alias{lookup.random.part.backref.estimable.unit}
\title{lookup possible information of ETA[i]}
\usage{
lookup.random.part.backref.estimable.unit(esu)
}
\arguments{
\item{esu}{est estimable unit}
}
\value{
NULL, or NAs or index
}
\description{
If available, find which the random part corresponding to original ETAs
}
