% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{lookup.random.part.backref.random.disturbation}
\alias{lookup.random.part.backref.random.disturbation}
\title{only for nonmem importing purpose
we may save hidden attribute backref referring to origianl index of the THETA, ETA's
this function extract those values}
\usage{
lookup.random.part.backref.random.disturbation(rd)
}
\arguments{
\item{rd}{random.disturbation}
}
\value{
the found references
}
\description{
only for nonmem importing purpose
we may save hidden attribute backref referring to origianl index of the THETA, ETA's
this function extract those values
}
