% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{match.symbolic}
\alias{match.symbolic}
\title{match.symbolic}
\usage{
match.symbolic(x, table, na = NA)
}
\arguments{
\item{x}{expression}

\item{table}{lookup table}

\item{na}{na symbol}
}
\value{
index found
}
\description{
match for symbolic expressions, similar to R itself's \code{match} for characters
here \code{x} is just expression, not having pattern variables, the equality test
just comparing them using R's \code{==} equality test operator
}
\details{
This is totally different from \code{symbolic.match}, should not be missly used
}
