% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympy_import.R
\name{Math.symbolic}
\alias{Math.symbolic}
\title{Math Functions for Symbolic Expressions}
\usage{
\method{Math}{symbolic}(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Further arguments passed to each method}
}
\value{
All these methods return a symbolic expression.
}
\description{
Math Functions for Symbolic Expressions
}
\examples{
syms_init(x, y, z)
sqrt(x)
}
