% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25plainmath.001matrix.R
\name{matrix.blockrize.permutation}
\alias{matrix.blockrize.permutation}
\title{find a permutation for better organization for blockrize
Here we assume \code{m} is symmetric}
\usage{
matrix.blockrize.permutation(m, eps = .Machine$double.eps * 10000)
}
\arguments{
\item{m}{matrix}
}
\value{
permutation
}
\description{
find a permutation for better organization for blockrize
Here we assume \code{m} is symmetric
}
