% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25plainmath.001matrix.R
\name{matrix.decompose.to.list}
\alias{matrix.decompose.to.list}
\title{inverse changes
change a matrix into block diagonal format, if possible
Not necessary to be symmetric, but need to be square}
\usage{
matrix.decompose.to.list(m0, eps = .Machine$double.eps * 10000,
  return.split.pt = FALSE)
}
\arguments{
\item{m0}{square matrix}

\item{eps}{threshold of zero}

\item{return.split.pt}{choose if return the list or matrix or just split points}
}
\value{
a list of diagonal blocks or split
}
\description{
inverse changes
change a matrix into block diagonal format, if possible
Not necessary to be symmetric, but need to be square
}
