% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25plainmath.001matrix.R
\name{matrix.direct.sum.list}
\alias{matrix.direct.sum.list}
\title{direct sum of linear mapping}
\usage{
matrix.direct.sum.list(l)
}
\arguments{
\item{l}{list of l}
}
\value{
diag block matrix
}
\description{
direct sum of linear mapping
}
