% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25plainmath.001matrix.R
\name{matrix.direct.sum.n}
\alias{matrix.direct.sum.n}
\title{handy interface
useful for combine block diagonal matrix to a full matrix}
\usage{
matrix.direct.sum.n(...)
}
\arguments{
\item{...}{any matrix}
}
\value{
matrix
}
\description{
handy interface
useful for combine block diagonal matrix to a full matrix
}
