% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25plainmath.001matrix.R
\name{matrix.direct.sum2}
\alias{matrix.direct.sum2}
\title{direct sum of linear mapping}
\usage{
matrix.direct.sum2(a = as.matrix(a), b = as.matrix(b))
}
\arguments{
\item{a}{matrix}

\item{b}{matrix}
}
\value{
diag block matrix a,b
}
\description{
direct sum of linear mapping
}
