% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.formula.fun.R
\name{monitor.formula.fun}
\alias{monitor.formula.fun}
\title{Default formula to monitor genetic evolution}
\usage{
monitor.formula.fun(obj)
}
\arguments{
\item{obj}{The solution object of the \code{GA} package.}
}
\description{
Default formula to monitor genetic evolution
}
\seealso{
genetic.search
}
