% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.002monomial.R
\name{monomial.of.theta.from.list.of.product}
\alias{monomial.of.theta.from.list.of.product}
\title{monomial.of.theta.from.list.of.product}
\usage{
monomial.of.theta.from.list.of.product(le)
}
\arguments{
\item{le}{list of product}
}
\value{
list of length 2, like \code{ [coefficients, monomial]}
}
\description{
similar to \code{\link{monomial.of.random.variable.from.list.of.product}} \cr
For a list of expressions, convert it as monomial of \code{eps(ETA[i])}'s\cr
\code{[ coefficients, THETA[i] ] ==> coefficients * ( THETA[i]^di .... ) }
}
\author{
Mango solutions
}
