% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{morphism.composition.2}
\alias{morphism.composition.2}
\title{simplify compose two morphisms
do not check if compatible}
\usage{
morphism.composition.2(m1, m2)
}
\arguments{
\item{m1}{morphism1}

\item{m2}{morphism2}
}
\value{
a new morphism as \code{ m1 o m2 }, i.e. first \code{m2} then followed by \code{m1}
}
\description{
simplify compose two morphisms
do not check if compatible
}
