% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{morphism.composition.list}
\alias{morphism.composition.list}
\title{other handy interface for Product of morphisms}
\usage{
morphism.composition.list(ml)
}
\arguments{
\item{ml}{morphism list}
}
\value{
morphism being the chained up morphisms in \code{ml}
}
\description{
other handy interface for Product of morphisms
}
