% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{morphism.composition.n}
\alias{morphism.composition.n}
\title{another handy interface for undetermined number of morphisms composition}
\usage{
morphism.composition.n(...)
}
\arguments{
\item{...}{morphisms}
}
\value{
morphism as the composition result
}
\description{
another handy interface for undetermined number of morphisms composition
}
