% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{morphism.concrete.instantiate}
\alias{morphism.concrete.instantiate}
\title{morphism instantiator}
\usage{
morphism.concrete.instantiate(m, parameter.pool)
}
\arguments{
\item{m}{morphism}

\item{parameter.pool}{to which symbols we instantiate it}
}
\value{
an expression or a set of expressions
}
\description{
morphism instantiator
}
