% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{morphism.concrete.new}
\alias{morphism.concrete.new}
\title{morphism constructor}
\usage{
morphism.concrete.new(type, ...)
}
\arguments{
\item{type}{type of morphism}
}
\value{
the morphism object
}
\description{
morphism constructor
}
