% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.110morphism.combinators.R
\name{morphism.curry.morphism}
\alias{morphism.curry.morphism}
\title{if the curryed one is a morphism
\code{ f : (x1,x2,x3) |-> f(x1,x2,x3) }
\code{ g : (x1,x2) |-> g(x1,x2) }
\code{f g} is then \code{ (x1,x2,x3,x4) |-> f(g(x1,x2),x3,x4) }}
\usage{
morphism.curry.morphism(m0, m1)
}
\arguments{
\item{m0}{morphism}

\item{m1}{morphism}
}
\value{
new morphism
}
\description{
if the curryed one is a morphism
\code{ f : (x1,x2,x3) |-> f(x1,x2,x3) }
\code{ g : (x1,x2) |-> g(x1,x2) }
\code{f g} is then \code{ (x1,x2,x3,x4) |-> f(g(x1,x2),x3,x4) }
}
