% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.110morphism.combinators.R
\name{morphism.curry.point}
\alias{morphism.curry.point}
\title{Currys
\code{ f : (x1,x2,x3) |-> f(x1,x2,x3) }
\code{f x1} is then \code{ (x2,x3) |-> f(x1,x2,x3) }}
\usage{
morphism.curry.point(m0, e0, ind = 1)
}
\arguments{
\item{m0}{main morphism}

\item{e0}{first argument}

\item{ind}{to replace which args}
}
\value{
morphism
}
\description{
Currys
\code{ f : (x1,x2,x3) |-> f(x1,x2,x3) }
\code{f x1} is then \code{ (x2,x3) |-> f(x1,x2,x3) }
}
