% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.110morphism.combinators.R
\name{morphism.defined.by.groupeffect}
\alias{morphism.defined.by.groupeffect}
\title{test if \code{e} has any pattern recognized by \code{analyse.group.effect}}
\usage{
morphism.defined.by.groupeffect(e)
}
\arguments{
\item{e}{expression}
}
\value{
morphism object if \code{e} has a known group effect
}
\description{
test if \code{e} has any pattern recognized by \code{analyse.group.effect}
}
