% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.080morphism.domain.R
\name{morphism.domain.one.dimensional}
\alias{morphism.domain.one.dimensional}
\title{create one dimensional interval description}
\usage{
morphism.domain.one.dimensional(a, b)
}
\arguments{
\item{a}{left end}

\item{b}{right end}
}
\value{
domain description expression
}
\description{
create one dimensional interval description
}
