% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.110morphism.combinators.R
\name{morphism.groupeffect.instantiator}
\alias{morphism.groupeffect.instantiator}
\title{for a morphism defined by groupeffect
instantiate it}
\usage{
morphism.groupeffect.instantiator(m, parameter.pool)
}
\arguments{
\item{m}{morphism}
}
\value{
expression or equations
}
\description{
for a morphism defined by groupeffect
instantiate it
}
