% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{morphism.unapply}
\alias{morphism.unapply}
\title{\code{unapply( x+y+z , (x,y))} will result the following morphism:
\code{(x,y) -> x+y +z}}
\usage{
morphism.unapply(e, tuple, ...)
}
\arguments{
\item{e}{expression}

\item{tuple}{the arguments}

\item{...}{other arguments passed into general.pointwise morphism constructor}
}
\value{
morphism
}
\description{
\code{unapply( x+y+z , (x,y))} will result the following morphism:
\code{(x,y) -> x+y +z}
}
