% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50pattern.002error.default.R
\name{multiple.dependent.variable.apply}
\alias{multiple.dependent.variable.apply}
\title{apply a multiple.dependent.variable pattern}
\usage{
multiple.dependent.variable.apply(m, ...)
}
\arguments{
\item{m}{multiple.dependent.variable object}
}
\value{
expression
}
\description{
apply a multiple.dependent.variable pattern
}
