% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{N.fixed.effects}
\alias{N.fixed.effects}
\title{return number of fixed in a disturbation, can be 0}
\usage{
N.fixed.effects(rd)
}
\arguments{
\item{rd}{random disturbation}
}
\value{
number of FIXed parameter
}
\description{
return number of fixed in a disturbation, can be 0
}
