% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{N.random.effects}
\alias{N.random.effects}
\title{return number of random effects in a disturbation}
\usage{
N.random.effects(rd)
}
\arguments{
\item{rd}{random disturbation}
}
\value{
number of RANDOM parameter
}
\description{
return number of random effects in a disturbation
}
