% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05graphic.000common.R
\name{neighbourhood.of.v}
\alias{neighbourhood.of.v}
\title{neighbourhood.of.v}
\usage{
neighbourhood.of.v(G, v1)
}
\arguments{
\item{G}{: \code{G} is a directed graph, which is represented as list of vertex and edges; edges are list of vertex pairs <from, to>}

\item{v1}{: vertex set}
}
\value{
list of incoming and outcoming list
}
\description{
For a vertex set \code{v1}, find the neighbourhood of \code{v1} in the directed Graph \code{G} \cr
return \code{v.income} and \code{v.outcome}, i.e. the incoming vertex to \code{v1} and output vertex from \code{v1}
}
\author{
jjxie
}
