% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{nonmem.eqns.to.r.eqns}
\alias{nonmem.eqns.to.r.eqns}
\title{batch converting from the NONMEM PK like code to r valid code}
\usage{
nonmem.eqns.to.r.eqns(eqns)
}
\arguments{
\item{eqns}{strings for nonmem equations}
}
\value{
equation as r expressions
}
\description{
some special change is applied that the following analysing code can run correctly
}
