% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{nonmem.parameters.specialarray.to.r.array}
\alias{nonmem.parameters.specialarray.to.r.array}
\title{nonmem.parameters.specialarray.to.r.array}
\usage{
nonmem.parameters.specialarray.to.r.array(e)
}
\arguments{
\item{e}{expression}
}
\value{
expression
}
\description{
simplifier generated by \code{\link{nonmem.parameters.specialarray.to.r.array.rules}}
}
