% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\docType{data}
\name{nonmem.parameters.specialarray.to.r.array.rules}
\alias{nonmem.parameters.specialarray.to.r.array.rules}
\title{nonmem.parameters.specialarray.to.r.array.rules}
\format{An object of class \code{list} of length 32.}
\usage{
nonmem.parameters.specialarray.to.r.array.rules
}
\description{
rules for changing \code{THETA(i),ETA(i)} to \code{THETA[i],ETA[i]}, \cr
both \code{EPS(i),ERR(i)} are changed to \code{EPS[i]} \cr
}
\details{
possible NONMEM(fortran) arrays are changed to R representations, including \cr
\code{A, DADT} etc.

NONMEM(fortran) 's function are changed to R equivlent, e.g. \code{LOG10(x)} to \code{log10(x)}

this is a rule definition
}
\seealso{
\code{\link{character.rules.to.list.rules}}, \code{\link{symbolic.simplify.gigo}}, \code{\link{nonmem.parameters.specialarray.to.r.array}}
}
\keyword{datasets}
