% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\docType{data}
\name{nonmem.r.expression.to.nonmem.rule}
\alias{nonmem.r.expression.to.nonmem.rule}
\title{nonmem.r.expression.to.nonmem.rule}
\format{An object of class \code{list} of length 12.}
\usage{
nonmem.r.expression.to.nonmem.rule
}
\description{
r expression to nonmem, still expression to expression, Array are kept as r Array
It's the inverse direct of \code{\link{nonmem.parameters.specialarray.to.r.array}}
}
\keyword{datasets}
