% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normailze.R
\name{normalize}
\alias{normalize}
\title{Normalize a dataset}
\usage{
normalize(X.df, custom.mins)
}
\arguments{
\item{X.df}{The input dataset to be normalized}

\item{custom.mins}{A list of user-defined minimum values for dataset columns.}
}
\value{
A list with two values:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    The input dataset normalized by subtracting the mean and dividing by the standard deviation.

    A list with some column statistic: std.min, mean, sd
}\if{html}{\out{</div>}}
}
\description{
Normalize a dataset
}
\examples{
\donttest{
# set-up a toy example dataset
x1<-runif(100, min=2, max=67)
x2<-runif(100, min=0.01, max=0.1)
X <- data.frame(x1=x1, x2=x2)
# set up a "true" non-linear relationship
# with some noise
y <- log10(x1^2*x2) + rnorm(100, 0, 0.001)
regressors <- c('x1','x2')
best.vars <- c('inv.mul.x1.x2','mul.x1.x2','x2')

transformations=list(
  "log"=function(rdf, x, stats){ log(x) },
  "log_x1_p"=function(rdf, x, stats){ log(rdf$x1 + x) },
  "inv"=function(rdf, x, stats){ 1/x }
)

# parse variables
parsed.vars <- symbolicr::parse.vars(best.vars, regressors, transformations)
# standardize
norm.res <- symbolicr::normalize(X, custom.mins=list())
X.std <- norm.res$X.std
X.mean.sd <-norm.res$mean.sd
# compute regressors
X.def <- symbolicr::regressors(X.std, parsed.vars,
                   transformations, X.mean.sd, regressors.min.values=NULL)
X.min.values <- X.def$min.values
formula.df <- X.def$regressors

# extract coefficients of given formula
dataset.std <- cbind(formula.df, y)
cur.formula.str <- paste0('y',"~",paste(best.vars, collapse=' + '))
base.lm <- lm(as.formula(cur.formula.str), data=formula.df)
base.lm$coefficients
}

}
