% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normailze.test.R
\name{normalize.test}
\alias{normalize.test}
\title{Normalize a dataset with pre-defined column mean and standard deviation}
\usage{
normalize.test(X.df, mean.sd)
}
\arguments{
\item{X.df}{The input dataset to be normalized}

\item{mean.sd}{list of user-defined mean and sd values for every dataset column.}
}
\value{
The input dataset normalized by subtracting the mean and dividing by the standard deviation.
}
\description{
Normalize a dataset with pre-defined column mean and standard deviation
}
