% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{number.of.free.parameters}
\alias{number.of.free.parameters}
\title{account the THETA's or THETA_HAT's in a list of expressions}
\usage{
number.of.free.parameters(le)
}
\description{
account the THETA's or THETA_HAT's in a list of expressions
}
