% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{ode.apropos}
\alias{ode.apropos}
\title{simply find a library definition}
\usage{
ode.apropos(keywords)
}
\arguments{
\item{keywords}{keywords for looking up}
}
\value{
the entry
}
\description{
simply find a library definition
}
