% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{ode.choose.nonmem.subroutine}
\alias{ode.choose.nonmem.subroutine}
\title{by some criterion(TO DO)
determine which ODE solver be used in NONMEM}
\usage{
ode.choose.nonmem.subroutine(ode)
}
\arguments{
\item{ode}{ode system}
}
\value{
statement
}
\description{
by some criterion(TO DO)
determine which ODE solver be used in NONMEM
}
