% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\docType{data}
\name{ode.librarys}
\alias{ode.librarys}
\title{will put more and more know ode equations here}
\format{An object of class \code{list} of length 1.}
\usage{
ode.librarys
}
\description{
will put more and more know ode equations here
}
\keyword{datasets}
