% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{ode.update.ijv}
\alias{ode.update.ijv}
\title{change \code{A[[i]][[j]] = v}}
\usage{
ode.update.ijv(ode, i, j, v)
}
\arguments{
\item{ode}{ode system}

\item{i}{x index}

\item{j}{y index}

\item{v}{new value}
}
\value{
new system
}
\description{
change \code{A[[i]][[j]] = v}
}
\examples{
 # tmp = matrix(1:4,2)
 # set ode to zero
 # ode.update.ijv(ode, row(tmp), col(rmp), 0)
}
