% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympy_import.R
\name{Ops.symbolic}
\alias{Ops.symbolic}
\title{Ops S3 Methods for Symbolic Expressions}
\usage{
\method{Ops}{symbolic}(e1, e2)
}
\arguments{
\item{e1, e2}{Left argument and right argument.}
}
\value{
All these methods return a symbolic expression.
}
\description{
Ops S3 Methods for Symbolic Expressions
}
\examples{
syms_init(x, y, z)
x + y
x * y
x ^ y
z != x + y
z != x + y | x + 1L == x + 1
}
