% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25plainmath.001matrix.R
\name{overall.dim.names.matrix.list}
\alias{overall.dim.names.matrix.list}
\title{get dim names from a matrix list}
\usage{
overall.dim.names.matrix.list(ml, prefix = "ETA")
}
\arguments{
\item{ml}{matrix list}

\item{prefix}{optional prefix}
}
\value{
names
}
\description{
get dim names from a matrix list
}
