% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{parse.if}
\alias{parse.if}
\title{parse.if}
\usage{
parse.if(txts)
}
\arguments{
\item{txts}{a character vector containing lines of NONMEM abbreviative code}
}
\value{
a list representing parsed structure,  \cr
         each entry in the list has a \code{type} entry(recording the type of this entry), \cr
         a \code{val} entry (recording a string which is R valid and parseable) \cr
         and a \code{pos} (the starting position in the original stream \code{txt}) entry
}
\description{
parse and convert NONMEM \code{if/then/else} (might be nested) Statements \cr
into valid R equivlent one line statement
}
\details{
multiple lines or \code{ELSEIF} will be changed into R's statements using braces \code{{}}
}
\author{
Mango solutions
}
