% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.vars.R
\name{parse.vars}
\alias{parse.vars}
\title{Parse text representation of a non-linear formula term}
\usage{
parse.vars(cur.vars, base.regressors, transformations = list())
}
\arguments{
\item{cur.vars}{An array of non-linear formula terms. Ex. \code{cur.vars <- c('a','mul.a.b')} represents the formula \code{y ~ a + a*b}}

\item{base.regressors}{The \code{names(X)} where X is the input dataset.}

\item{transformations}{A list of potentially non-linear transformations allowed in \code{cur.vars}.}
}
\value{
A list with the base regressors and the transformation function to be applied
}
\description{
Return a list with the non-linear function applied and the base terms of the formula.
}
\examples{
parse.vars(
 c('log.a', 'mul.a.mul.b.c'),
 base.regressors=c('a','b','c'),
 transformations=list('log'=log)
)

}
\seealso{
random.search

genetic.search

serialize.vars
}
