% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{parse1}
\alias{parse1}
\title{parse1
parse oneline text to a expression (call)}
\usage{
parse1(s)
}
\arguments{
\item{s}{oneline string}
}
\value{
an R call
}
\description{
parse1
parse oneline text to a expression (call)
}
\author{
jjxie
}
