% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{pattern.affine.function.of.theta}
\alias{pattern.affine.function.of.theta}
\title{match \code{Sum c_i * THETA[i]}}
\usage{
pattern.affine.function.of.theta(e)
}
\arguments{
\item{e}{expression}
}
\value{
row matrix and intercept
}
\description{
match \code{Sum c_i * THETA[i]}
}
