% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{pattern.affine.function.theta.or.ifelse}
\alias{pattern.affine.function.theta.or.ifelse}
\title{pattern.affine.function.theta.or.ifelse}
\usage{
pattern.affine.function.theta.or.ifelse(e0)
}
\arguments{
\item{e0}{expression}
}
\value{
if matched, a \code{Sum.of.Morphisms} morphism
}
\description{
similar as above \code{\link{pattern.product.expression.containing.ifelse}}
should be a morphism level of \code{\link{expand.as.sum.of.product}} .
For example, \code{ S1 + S2 + S3 } will be decomposed, 
if those expression \code{Si} is matched by 
\code{\link{pattern.product.expression.containing.ifelse}}.
}
