% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{pattern.affine.mapping.factor.effect}
\alias{pattern.affine.mapping.factor.effect}
\title{pattern.affine.mapping.factor.effect}
\usage{
pattern.affine.mapping.factor.effect(e0)
}
\arguments{
\item{e0}{expression}
}
\value{
a \code{affine.mapping.factor.effect} type object which contain both the levels and covariate name and the morphism(affine mapping) of expression corresponding to levels
}
\description{
match a the following pattern 
\code{ifelse(var1,..., ifelse(var1,...,ifelse))}
and which can be converted into affine mapping
}
