% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{pattern.affine.mapping.of.theta}
\alias{pattern.affine.mapping.of.theta}
\title{multiple expressions}
\usage{
pattern.affine.mapping.of.theta(mes)
}
\arguments{
\item{mes}{multiple expressions as a vector mapping}
}
\value{
matrix and right vector if the whole mapping is an affine mapping
}
\description{
multiple expressions
}
