% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.00math.R
\name{pattern.binomial.combination}
\alias{pattern.binomial.combination}
\title{binomial
assume \code{X ~ binom(Q)}
\code{P(X=X1) = Q1}
\code{P(X=X2) = 1-Q1}
then \code{E(X) = Q1 * X1 + (1-Q1) * X2}
i.e. \code{E(X)} (binomial combination) is the expection of a bernoulli distribution}
\usage{
pattern.binomial.combination(e)
}
\arguments{
\item{e}{expression}
}
\value{
if matched, the combination is viewed as a random variable, and return the discription of the random variable
}
\description{
binomial
assume \code{X ~ binom(Q)}
\code{P(X=X1) = Q1}
\code{P(X=X2) = 1-Q1}
then \code{E(X) = Q1 * X1 + (1-Q1) * X2}
i.e. \code{E(X)} (binomial combination) is the expection of a bernoulli distribution
}
