% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.00math.R
\name{pattern.convex.combination}
\alias{pattern.convex.combination}
\title{similar to bernoulli 
see \code{\link{pattern.binomial.combination}}
however, \code{Xi} is now thought like a multinomial distribution}
\usage{
pattern.convex.combination(e)
}
\arguments{
\item{e}{expression}
}
\value{
\code{NULL} if not matched, random variable's distribution if matched
}
\description{
similar to bernoulli 
see \code{\link{pattern.binomial.combination}}
however, \code{Xi} is now thought like a multinomial distribution
}
