% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.002monomial.R
\name{pattern.distrib.fixed}
\alias{pattern.distrib.fixed}
\title{pattern.distrib.fixed}
\usage{
pattern.distrib.fixed(e)
}
\arguments{
\item{e}{expression}
}
\value{
mean and vars if matched; else \code{NULL}
}
\description{
match a expression which not having ETA's
}
