% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.002monomial.R
\name{pattern.distrib.lognormal}
\alias{pattern.distrib.lognormal}
\title{pattern.distrib.lognormal}
\usage{
pattern.distrib.lognormal(e, ...)
}
\arguments{
\item{e}{expression}

\item{...}{options passed to \code{\link{pattern.distrib.normal}}}
}
\value{
if matched, return the distribution; else \code{NULL}
}
\description{
similar to \code{\link{pattern.distrib.normal}} \cr
For \code{ A = B * exp(ETA[i])} , \cr
extract means of \code{log(A)} is \code{log(B)} \cr
and  covariance of \code{log(A)} is \code{Cov(ETA[i])}
}
\author{
Mango solutions
}
