% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.002monomial.R
\name{pattern.distrib.normal}
\alias{pattern.distrib.normal}
\title{pattern.distrib.normal}
\usage{
pattern.distrib.normal(e, ignore.non.atomic.coefficient = FALSE)
}
\arguments{
\item{e}{expression}

\item{ignore.non.atomic.coefficient}{do not match if the coefficient random varible is non-atomic}
}
\value{
if matched, a list indicating the distribution, including mean and covariance; if not matched then \code{NULL}
}
\description{
used by higher level pattern matcher, if not matched, then try next distribution pattern
}
\details{
try to match a expression to normal distribution, e.g.\cr
\code{ expressions + coeffs*ETA[i] } \cr
\code{ expressions + coeffs*EPS[i] }

the \code{\link{create.polynomial.of.random.variable}} are used here for combining \code{ETA[i]}'s or \code{EPS[i]}'s , for example \cr
\code{ TVCL + ETA[1] + 100 * ETA[1] + TVCL*ETA[1] } should be coverted to \code{ TVCL + ( 101 + TVCL)*ETA[1] } \cr
and then mean is \code{TVCL} and covariance is \code{(101+TVCL)^2 * Cov(ETA[1])}  \cr
}
\author{
Mango solutions
}
