% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50pattern.002error.default.R
\name{pattern.error.default}
\alias{pattern.error.default}
\title{pattern.error.default}
\usage{
pattern.error.default(eqns = NULL, md = NULL)
}
\arguments{
\item{eqns}{equations in \code{$ERROR} block}

\item{md}{\code{importNmMod} model}
}
\value{
list of definition of observation and \code{IPRED}
}
\description{
turn set of equations define Error block into errorObjs
}
\author{
Mango solutions
}
