% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{pattern.expression.simple.assignment}
\alias{pattern.expression.simple.assignment}
\title{match a simple assignment pattern}
\usage{
pattern.expression.simple.assignment(e)
}
\arguments{
\item{e}{expression}
}
\value{
lhs and rhs
}
\description{
match a simple assignment pattern
}
