% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{pattern.factorizable.one.ifelse}
\alias{pattern.factorizable.one.ifelse}
\title{match a simpliest pattern
\code{ifelse(var1,..., ifelse(var1,...,ifelse))}}
\usage{
pattern.factorizable.one.ifelse(e0)
}
\description{
match a simpliest pattern
\code{ifelse(var1,..., ifelse(var1,...,ifelse))}
}
