% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.001analyse.R
\name{pattern.linear.by.binomial.categorical}
\alias{pattern.linear.by.binomial.categorical}
\title{match a convex combination of two similar linear pattern
\code{ p[1],p[2],p[3],p[4] -> grpvar[1] * (p[1] + grpvar[2]*p[2]) + (1-grpvar[1]) * (p[3] + grpvar[2]*p[4]) }}
\usage{
pattern.linear.by.binomial.categorical(e)
}
\arguments{
\item{e}{expression}
}
\value{
pattern description
}
\description{
match a convex combination of two similar linear pattern
\code{ p[1],p[2],p[3],p[4] -> grpvar[1] * (p[1] + grpvar[2]*p[2]) + (1-grpvar[1]) * (p[3] + grpvar[2]*p[4]) }
}
