% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50pattern.002error.default.R
\name{pattern.multiple.dimensional.by.factor}
\alias{pattern.multiple.dimensional.by.factor}
\title{multiple DVs
try to match following:
\code{ ifelse( DVID == 2, expression1, ifelse(DVID <=1 , expression2, NA)) }}
\usage{
pattern.multiple.dimensional.by.factor(e, varname = quote(DVID),
  var.levels = as.character(0:2))
}
\arguments{
\item{e}{expression}

\item{varname}{variable name}

\item{var.levels}{possible levels}
}
\value{
list of expressions by var.levels
}
\description{
multiple DVs
try to match following:
\code{ ifelse( DVID == 2, expression1, ifelse(DVID <=1 , expression2, NA)) }
}
