% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50pattern.001pk.default.R
\name{pattern.pk.default}
\alias{pattern.pk.default}
\title{pattern.pk.default}
\usage{
pattern.pk.default(eqns = NULL, md = NULL, L = NULL)
}
\arguments{
\item{eqns}{equations}

\item{md}{model returned by \code{importNmMod} , if \code{NULL}, the covariance are left as is, not as numeric values}

\item{L}{the analysed result by \code{analyse.0.0}, override the \code{md} and \code{eqns}}
}
\value{
list of definition of PK parameters, including their distributions, how the mean (including group or covariant effect) and variance are defined, e.t.c
}
\description{
Try to match typical control file sytle if we're using NONMEM PREDPP subroutines \cr
Assume control stream is PK parameters orientated, i.e. we know the whold model given we know definitions of all PK parameters
}
\details{
For each connected component return by \code{\link{analyse.0.0}}, check if it can be explained by
simple relations as additive, multiplicative, affine, linear or binomial.linear combination of parameters and covariates
}
\author{
Mango solutions
}
