% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{pattern.product.expression.containing.ifelse}
\alias{pattern.product.expression.containing.ifelse}
\title{we do not want to make things too complex, hence no nested will be detected
Only those like 
\code{ ifelse(,,ifelse(,,ifelse(,,NA))) * something * THETA[i] } will de decomposed}
\usage{
pattern.product.expression.containing.ifelse(e0)
}
\arguments{
\item{e0}{expression}
}
\value{
if matched, type \code{Product.of.Morphisms} morphism
}
\description{
we do not want to make things too complex, hence no nested will be detected
Only those like 
\code{ ifelse(,,ifelse(,,ifelse(,,NA))) * something * THETA[i] } will de decomposed
}
