% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe.r.squared.formula.len.fitness.R
\name{pe.r.squared.formula.len.fitness}
\alias{pe.r.squared.formula.len.fitness}
\title{Default fitness function for \code{genetic.search}}
\usage{
pe.r.squared.formula.len.fitness(errs.m, max.formula.len)
}
\arguments{
\item{errs.m}{a 1-row data.frame with cross-validation results}

\item{max.formula.len}{the maximum length of the formula we are searching for}
}
\value{
A single double value, the bigger the value the better the formula.
}
\description{
Default fitness function for \code{genetic.search}
}
\examples{
\donttest{
# set-up a toy example dataset
x1<-runif(100, min=2, max=67)
x2<-runif(100, min=0.01, max=0.1)
X <- data.frame(x1=x1, x2=x2)
# set up a "true" non-linear relationship
# with some noise
y <- log10(x1^2*x2) + rnorm(100, 0, 0.001)

genetic.search(
 X,
 y,
 n.squares=1,
 maxiter=10,
 glob.filepath=NULL,
 local.filepath=NULL,
 memoization=FALSE,
 pcrossover=0.2,
 pmutation=0.8,
 seed=NULL,
 max.formula.len=2,
 keepBest=TRUE,
 K=2,
 N=3,
 popSize = 5,
 fitness.fun=pe.r.squared.formula.len.fitness,
 best.vars.l = list(
       c('x1','x2')
 )
)
}
}
