% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{phase.space.dim}
\alias{phase.space.dim}
\title{the length of state.vector
the dimensional of state space (or say, phase space)}
\usage{
phase.space.dim(ode)
}
\arguments{
\item{ode}{the ode equations}
}
\value{
number
}
\description{
the length of state.vector
the dimensional of state space (or say, phase space)
}
