% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprint.R
\name{pprint}
\alias{pprint}
\title{Pretty Print of Symbolic Expressions}
\usage{
pprint(expr, ...)
}
\arguments{
\item{expr}{A symbolic expression}

\item{...}{Printing options, there are six options available in sympy, namely
"wrap_line", "num_columns", "use_unicode", "full_prec", "order" and
"use_unicode_sqrt_char". Detail see the documentation at ........}
}
\value{
It invisibly returns the expression.
}
\description{
Pretty Print of Symbolic Expressions
}
