% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pred.vs.obs}
\alias{pred.vs.obs}
\title{Plot "predicted vs observed" using a given formula}
\usage{
pred.vs.obs(
  complete.X.df,
  y,
  cur.vars,
  custom.abs.mins,
  K,
  N,
  transformations = list(log10 = function(rdf, x, z) {
     log10(0.1 + abs(z$min) + x)

    }, inv = function(rdf, x, z) {
     1/(0.1 + abs(z$min) + x)
 }),
  cv.norm = TRUE,
  errors.x = 3.2,
  errors.y = 5,
  with.names = FALSE
)
}
\arguments{
\item{complete.X.df}{The dataset to be used for the test}

\item{y}{The independent variable}

\item{cur.vars}{An array of non-linear formula terms to be tested. \code{cur.vars <- c('a','mul.a.b')} will test the formula \code{y ~ a + a*b}}

\item{custom.abs.mins}{A list of user-defined minimum values for dataset columns.}

\item{K}{The number of parts the dataset is split into for K-fold cross-validation.}

\item{N}{The number of times the K-fold validation is repeated, shuffling the dataset row orders before each time.}

\item{transformations}{A list of potentially non-linear transformations allowed in \code{cur.vars}.}

\item{cv.norm}{Normalize regressors after train-validation split in inner cross-validation loop.}

\item{errors.x}{position of R^2 and PE in plot (x axis)}

\item{errors.y}{position of R^2 and PE in plot (y axis)}

\item{with.names}{Plot also product name}
}
\value{
ggplot2::ggplot object ready to be plotted
}
\description{
Predicted values are obtained out-of-sample using K-fold cross-validation repeated N times and averaged.
}
\examples{
\donttest{
# set-up a toy example dataset
x1<-runif(100, min=2, max=67)
x2<-runif(100, min=0.01, max=0.1)
X <- data.frame(x1=x1, x2=x2)
# set up a "true" non-linear relationship
# with some noise
y <- log10(x1^2*x2) + rnorm(100, 0, 0.001)
n.squares=1
K=2
N=2
seed=1001

transformations=list(
  "log"=function(rdf, x, stats){ log(x) },
  "log_x1_p"=function(rdf, x, stats){ log(rdf$x1 + x) },
  "inv"=function(rdf, x, stats){ 1/x }
)

best.f <- list(
  c('inv.mul.x1.x2','x1'),
  c('inv.mul.x1.x2','x2')
)
# analyze different formulas with predicted vs observed plot.
# the more data points align round the x=y line, the more
# the prediction of the corresponding formula is accurate
plots <- lapply(best.f, function(test.f){
    pred.vs.obs(X,
                y, test.f,
                list(), K, N,
                transformations,
                cv.norm = FALSE,
                errors.x=2.5, errors.y=0.5
   )
})
}
}
