% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{print.affine.mapping.factor.effect}
\alias{print.affine.mapping.factor.effect}
\title{the S3 version}
\usage{
\method{print}{affine.mapping.factor.effect}(factors.affine)
}
\description{
the S3 version
}
