% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.005ifelse.factor.R
\name{print.affine.mapping0}
\alias{print.affine.mapping0}
\title{pretty printer for the affine mapping
the working horse}
\usage{
\method{print}{affine.mapping0}(mapping)
}
\description{
pretty printer for the affine mapping
the working horse
}
