% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{print.distribution.common0}
\alias{print.distribution.common0}
\title{printer working horse}
\usage{
\method{print}{distribution.common0}(d)
}
\arguments{
\item{d}{distribution}
}
\description{
printer working horse
}
