% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{print.listoflist}
\alias{print.listoflist}
\title{print.listoflist}
\usage{
\method{print}{listoflist}(x)
}
\arguments{
\item{x}{list of list of expressions}
}
\value{
string
}
\description{
Just for debuging purpose, pretty printer for list of list\cr
Will display list of list as \code{[ [A11, A12, A13, ... ] , [A21, A22, A23, ... ], ... ]}
}
\author{
jjxie
}
