% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{print.morphism}
\alias{print.morphism}
\title{will deal with Composition special
printing additional information of the components morphism
affine mapping will also show the matrix
the structure will be shown more clearer}
\usage{
\method{print}{morphism}(m)
}
\arguments{
\item{m}{morphism}
}
\description{
will deal with Composition special
printing additional information of the components morphism
affine mapping will also show the matrix
the structure will be shown more clearer
}
