% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.100morphism.R
\name{print.morphism0}
\alias{print.morphism0}
\title{FORMATTER
simple pretty printer for morphism}
\usage{
\method{print}{morphism0}(m, WW = 60)
}
\arguments{
\item{m}{morphism}
}
\value{
string
}
\description{
FORMATTER
simple pretty printer for morphism
}
