% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50pattern.002error.default.R
\name{print.multiple.dependent.variable}
\alias{print.multiple.dependent.variable}
\title{S3 print function}
\usage{
\method{print}{multiple.dependent.variable}(m0)
}
\description{
S3 print function
}
