% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{print.random.disturbation}
\alias{print.random.disturbation}
\title{S3 version printer for disturbation}
\usage{
\method{print}{random.disturbation}(d)
}
\arguments{
\item{d}{disturbation}
}
\description{
S3 version printer for disturbation
}
