% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{QCONS}
\alias{QCONS}
\title{QCONS : quoted CONS}
\usage{
QCONS(x1, ...)
}
\arguments{
\item{x1}{: the function, can be a variable, or quoted symbol}

\item{...}{: other arguments}
}
\value{
the call
}
\description{
Construct an R call, the first argument is evaluated, the others are quoted
}
\details{
\code{QCONS('*', 1, 2+3)} \samp{->} \code{1 * (2+3)}

this is useful , because

e.g.   \code{QCONS(is.integer ,a )} \samp{->} \code{is.integer(a)}

where \code{a} is a variable

then we can use \code{eval( QCONS(is.integer, a) )} to the result
}
\author{
jjxie
}
