% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{R}
\alias{R}
\title{Convert a tuple or dict to R object}
\usage{
R(x, add_sym_class = TRUE)
}
\arguments{
\item{x}{A python object}

\item{add_sym_class}{Logical value indicating whether to add "symbolic" class to the elements
that can not be converted.}
}
\value{
A corresponding R object, if any of its elements can not be converted to
    R, a "symbolic" class will be added to that element.
}
\description{
Convert a tuple or dict to R object
}
\examples{
syms_init(x, y)
sym_func("div")(x^3L + 6L*x + 1L, x)
sym_func("div")(x^3L + 6L*x + 1L, x) \%>\% R
sym_func("div")(x^3L + 6L*x + 1L, x) \%>\% R \%>\% lapply(class)
}
