% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{random.disturbation.extract.fixed.domain}
\alias{random.disturbation.extract.fixed.domain}
\title{extract the domain of the fixed parameters}
\usage{
random.disturbation.extract.fixed.domain(d)
}
\arguments{
\item{d}{disturbation}
}
\value{
if fixed domain found, return domain expression
}
\description{
extract the domain of the fixed parameters
}
