% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{random.disturbation.new.eps}
\alias{random.disturbation.new.eps}
\title{generate a disturbation from expression containing EPS
non fixed
EPS[i] as random
extra backref}
\usage{
random.disturbation.new.eps(e)
}
\arguments{
\item{e}{expression}
}
\value{
disturbation
}
\description{
generate a disturbation from expression containing EPS
non fixed
EPS[i] as random
extra backref
}
