% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{random.disturbation.new.from.theta.eta}
\alias{random.disturbation.new.from.theta.eta}
\title{assume THETA[i] are FIXed and ETA[i] are RANDOM}
\usage{
random.disturbation.new.from.theta.eta(e)
}
\arguments{
\item{e}{expression}
}
\value{
random disturbation
}
\description{
assume THETA[i] are FIXed and ETA[i] are RANDOM
}
