% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{random.disturbation.new}
\alias{random.disturbation.new}
\title{construct a disturbation
This is a handy interface for defining your own distribution class}
\usage{
random.disturbation.new(e, fixed, random, distribution)
}
\arguments{
\item{e}{expression}

\item{fixed}{fixed variables in expression}

\item{random}{random variables in expression}

\item{distribution}{distribution for random vectors}
}
\value{
a disturbation object which can be added to some variable
}
\description{
\code{ FIXED[1] * exp( RANDOM[1] ) }
The code will try to unapply it and determine the index of the fixed and random part, also the dimension of both
part. And distribution should contain the information of random part
}
\details{
Of course, we are losing the original information on the combinated parameter - the whole one, e.g.
\code{ a * exp( ETA[1] ) } , we are not recognizing it as a whole is log normal now
}
