% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{random.disturbation.update.fixed.domain}
\alias{random.disturbation.update.fixed.domain}
\title{update the domain part of the fixed parameters}
\usage{
random.disturbation.update.fixed.domain(d, new.domain)
}
\arguments{
\item{d}{disturbation}

\item{new.domain}{new domain expression}
}
\description{
update the domain part of the fixed parameters
}
