% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{py_available}
\alias{reexports}
\alias{py_numpy_available}
\alias{reexports}
\alias{py_module_available}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{reticulate}{\code{\link[reticulate]{py_available}}, \code{\link[reticulate]{py_numpy_available}}, \code{\link[reticulate]{py_module_available}}}
}}

