% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.regressors.names.R
\name{regressors.names}
\alias{regressors.names}
\title{Compute all possible formula terms from a given dataset}
\usage{
regressors.names(complete.X.df, n.squares, transformations)
}
\arguments{
\item{complete.X.df}{The dataset that contains the base variables the formula is composed of (column-wise)}

\item{n.squares}{The maximum order of the polynomial composition of base variables. Ex. \verb{order 0 = a}, \verb{order 1 = a*b}, \verb{order 2 = a*b*c}}

\item{transformations}{A list of potentially non-linear transformations that can be applied on top of the squares. Ex. \verb{order 0, transformation=log10 = log10.a}. Input values are x and z, the array of numbers to be transformed (training set only), and the min max statistics (on the global dataset) respectively.}
}
\value{
An array of character with all the generated names for the generated variables
}
\description{
Compute all possible formula terms from a given dataset
}
\examples{
\donttest{
# set-up a toy example dataset
x1<-runif(100, min=2, max=67)
x2<-runif(100, min=0.01, max=0.1)
X <- data.frame(x1=x1, x2=x2)

n.squares=2

transformations=list(
  "log"=function(rdf, x, stats){ log(x) },
  "log_x1_p"=function(rdf, x, stats){ log(rdf$x1 + x) },
  "inv"=function(rdf, x, stats){ 1/x }
)

regressors <- c('x1','x2')

regressors.names(X, n.squares, transformations)
}
}
