% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.regressors.R
\name{regressors}
\alias{regressors}
\title{Compute regressors of a given non-linear formula}
\usage{
regressors(
  base.X.df.std,
  parsed.vars,
  transformations,
  X.mean.sd,
  regressors.min.values = NULL
)
}
\arguments{
\item{base.X.df.std}{A data.frame with the base variables that compose the terms of the non-linear formula}

\item{parsed.vars}{The results of the symbolicr::parsed.vars}

\item{transformations}{A list of potentially non-linear transformations that can be used in the formula terms.}

\item{X.mean.sd}{The \code{mean.sd} list entry of the returned value of \code{normalize}}

\item{regressors.min.values}{The output of \code{dataset.min.maxs}. If NULL, minimum values are computed on \code{base.X.df.std}. Use this parameter if you want to fix the minimum values, for instance if you compute it in a subset of the full dataset, as happens in the cross-validation.}
}
\value{
Return a data.frame with the columns corresponding to each formula term.
}
\description{
Compute regressors of a given non-linear formula
}
\examples{
\donttest{
# set-up a toy example dataset
x1<-runif(100, min=2, max=67)
x2<-runif(100, min=0.01, max=0.1)
X <- data.frame(x1=x1, x2=x2)
# set up a "true" non-linear relationship
# with some noise
y <- log10(x1^2*x2) + rnorm(100, 0, 0.001)
regressors <- c('x1','x2')
best.vars <- c('inv.mul.x1.x2','mul.x1.x2','x2')

transformations=list(
  "log"=function(rdf, x, stats){ log(x) },
  "log_x1_p"=function(rdf, x, stats){ log(rdf$x1 + x) },
  "inv"=function(rdf, x, stats){ 1/x }
)

# parse variables
parsed.vars <- symbolicr::parse.vars(best.vars, regressors, transformations)
# standardize
norm.res <- symbolicr::normalize(X, custom.mins=list())
X.std <- norm.res$X.std
X.mean.sd <-norm.res$mean.sd
# compute regressors
symbolicr::regressors(X.std, parsed.vars,
                   transformations, X.mean.sd, regressors.min.values=NULL)
}

}
\seealso{
random.search

genetic.search
}
