% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{remove.covariance.of.estimable.unit.estimable.model}
\alias{remove.covariance.of.estimable.unit.estimable.model}
\title{remove covariance relations for inner random variable of \code{estimable.unit}s in an \code{estimable.model}}
\usage{
remove.covariance.of.estimable.unit.estimable.model(esm, v1, v2,
  v1.subind = 1, v2.subind = 1)
}
\arguments{
\item{esm}{estimable model}

\item{v1}{estimable.unit 1}

\item{v2}{estimable.unit 2}

\item{v1.subind}{one estimable.unit may have more than one random varible in it, this indicate which variable are operated on}

\item{v2.subind}{similar to \code{v1.subind}}
}
\value{
the updated estimable.model
}
\description{
remove covariance relations for inner random variable of \code{estimable.unit}s in an \code{estimable.model}
}
\note{
scaler version, each argument has to be a scalar.
}
