% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{remove.one.line.code.estimable.model}
\alias{remove.one.line.code.estimable.model}
\title{remove one line code in the model spec}
\usage{
remove.one.line.code.estimable.model(esm, lhs, which.part = "PK.PROGRAM")
}
\arguments{
\item{esm}{estimable.model}

\item{lhs}{integer or character or quoted expression}
}
\value{
updated estimable.model
}
\description{
remove one line code in the model spec
}
