% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{remove.random.effect.estimable.unit}
\alias{remove.random.effect.estimable.unit}
\title{convert estimable unit to a simpler one without random effect}
\usage{
remove.random.effect.estimable.unit(m)
}
\arguments{
\item{m}{estimable.unit}
}
\value{
estimable.unit with random effect removed, or simply an expression
}
\description{
Atomic or Constant morphism are simplified into expreesion or atomic,
}
