% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{remove.random.effects.estimable.model}
\alias{remove.random.effects.estimable.model}
\title{Batch removing the random effects in an \code{estimable.model}}
\usage{
remove.random.effects.estimable.model(esm, varnames, all = FALSE)
}
\arguments{
\item{esm}{estimable.model}

\item{varnames}{variable names}
}
\value{
new model
}
\description{
Remember the function \code{\link{remove.random.effect.estimable.unit}} can specialize an \code{estimable.unit} into simpler one without random effect.
But it's boring to do it in a \code{for} loop by usr, to update any possible \code{estimable.unit} in a model.
So we provide this API for the purpose. The usage is simple, put the names of the \code{estimable.unit} in argument \code{varnames},
}
