% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50pattern.001pk.default.R
\name{reqns.to.nonmem.strings}
\alias{reqns.to.nonmem.strings}
\title{simplify output equations to NONMEM string}
\usage{
reqns.to.nonmem.strings(eqns)
}
\arguments{
\item{eqns}{equations}
}
\value{
one string
}
\description{
simplify output equations to NONMEM string
}
