% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{resolve.to.mapping.chain}
\alias{resolve.to.mapping.chain}
\title{resolve.to.mapping.chain}
\usage{
resolve.to.mapping.chain(eqns)
}
\arguments{
\item{eqns}{NONMEM equations}
}
\value{
list representing the discovered sections
}
\description{
By Graph of Symbol's, we can section up the equations into several stages\cr
convert equations to mappings' chain
}
\details{
This is done by using topological sort on the graph, it can discover very basic,\cr
rough sections in your PK/PRED code.
}
