% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{rexpression.to.nonmem.expression}
\alias{rexpression.to.nonmem.expression}
\title{rexpression.to.nonmem.expression}
\usage{
rexpression.to.nonmem.expression(e)
}
\arguments{
\item{e}{expression}
}
\value{
e expression
}
\description{
generated by \code{\link{nonmem.r.expression.to.nonmem.rule}}
}
