% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.001nonmem.R
\name{rexpression.to.string.nonmem}
\alias{rexpression.to.string.nonmem}
\title{rexpression.to.string.nonmem}
\usage{
rexpression.to.string.nonmem(e0)
}
\arguments{
\item{e0}{R expression}
}
\value{
a single string including <CR>
}
\description{
R expression to strings which is valid in NONMEM control stream \cr
\itemize{
 \item logical e.g. \code{ .EQ.  => == }
 \item braces \code{ {} } removed
 \item if-else re-organized as NONMEM control stream
 \item ARRAY \code{[]} is changed into \code{()}
 \item exponential \code{^} is changed into \code{**}
 \item redundant parentheses are removed, e.g. \code{ a + (b*c) } \samp{->} \code{ a + b*c }
}
}
