% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{runUnit.symbolicR}
\alias{runUnit.symbolicR}
\title{Executes the RUnit unit tests suite}
\usage{
runUnit.symbolicR(testPath = system.file(package = "symbolicR", "testing"),
  protocolFile = "symbolicRtestResults.html")
}
\arguments{
\item{testPath}{A single string holding the path to the internal test suite scripts}

\item{protocolFile}{Single string with the name of the HTML report file produced by RUnit for the internal unit tests}
}
\value{
standard internal test suite (object of class RUnitTestData returned by \code{runTestSuite} from the \code{RUnit} package)
}
\description{
contains unit tests distributed with the package
}
\examples{
\dontrun{ 
     runDDMoReMDLTests
}
}
\author{
Mango Solutions
}
\keyword{debugging}
\keyword{programming}
