% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{score.keywords.match}
\alias{score.keywords.match}
\title{Compute the matching score for keywords in a list of keywords}
\usage{
score.keywords.match(keywords, keywords.list)
}
\arguments{
\item{list}{of keywords}
}
\description{
Compute the matching score for keywords in a list of keywords
}
