% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.vars.R
\name{serialize.vars}
\alias{serialize.vars}
\title{Parse text representation of a non-linear formula term}
\usage{
serialize.vars(parsed.vars, base.regressors, transformations = list())
}
\arguments{
\item{parsed.vars}{A list with the base regressors and the transformation function to be applied}

\item{base.regressors}{The \code{names(X)} where X is the input dataset.}

\item{transformations}{A list of potentially non-linear transformations allowed in \code{cur.vars}.}
}
\value{
An array of serialized non-linear formula terms. Ex. \code{cur.vars <- c('a','mul.a.b')} represents the formula \code{y ~ a + a*b}
}
\description{
Return a list with the non-linear function applied and the base terms of the formula.
}
\examples{
serialize.vars(
 list(
   list("transformation.name"="log", "prods"=c("a")), 
   list("transformation.name"=NULL, "prods"=c("a","b","c"))
 ),
 base.regressors=c('a','b','c'),
 transformations=list('log'=log)
)

}
\seealso{
random.search

genetic.search

parse.vars
}
