% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{simplify.03}
\alias{simplify.03}
\title{simplify.03}
\usage{
simplify.03(e)
}
\arguments{
\item{e}{expression}
}
\value{
simplified expression
}
\description{
helper simplifier for following purpose \cr
 \code{(a^i)^j} \code{=>} \code{a^(i*j)} \cr
This simplifier is generated by \code{symbolic.simplify.gigo} factory, hence, also an example for using \cr
 \code{\link{symbolic.simplify.gigo}}
}
\author{
jjxie
}
