% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{simplify.1}
\alias{simplify.1}
\title{simplify.1}
\usage{
simplify.1(e)
}
\arguments{
\item{e}{expression}
}
\value{
simplified expression
}
\description{
Simplify level 1: one level higher than simplify, by expand an expression to list of list \cr
removing redundant terms, and combine back \cr
e.g.    \code{A*B + C - A*B -> C}
}
\author{
jjxie
}
