% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{simplify.2}
\alias{simplify.2}
\title{simplify.2
math level simplifier \cr}
\usage{
simplify.2(e0)
}
\arguments{
\item{e0}{R expression}
}
\value{
simplified expression
}
\description{
\code{log(A*B)} \samp{->} \code{log(A)+log(B)} \cr
     \code{log(A^B)} \samp{->} \code{ B * log(A) } \cr
     \code{exp(A)^B} \samp{->} \code{ exp(A*B) } \cr
     \code{exp(log(A))} \samp{->} \code{A} \cr
     \code{log(exp(A))} \samp{->} \code{A} \cr
     evaluate \code{log(atomic)} to a number
}
\seealso{
\code{\link{simplify}}
}
\author{
jjxie
}
