% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.080morphism.domain.R
\name{simplify.cartesian.product}
\alias{simplify.cartesian.product}
\title{e.g. RealNumber^1 * RealNumber^1 => RealNumber^2}
\usage{
simplify.cartesian.product(e)
}
\description{
e.g. RealNumber^1 * RealNumber^1 => RealNumber^2
}
