% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.001analyse.R
\name{simplify.collect.log.exp}
\alias{simplify.collect.log.exp}
\title{simplify.collect.log.exp
convert \code{ exp(log(A) + C*log(B)) } to \code{ A * B^C }}
\usage{
simplify.collect.log.exp(e)
}
\arguments{
\item{e}{expression}
}
\description{
simplify.collect.log.exp
convert \code{ exp(log(A) + C*log(B)) } to \code{ A * B^C }
}
