% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.001analyse.R
\docType{data}
\name{simplify.collect.log.exp.rules}
\alias{simplify.collect.log.exp.rules}
\title{reverse direction of \code{simplify.2} for \code{log} and \code{exp} expressions}
\format{An object of class \code{list} of length 6.}
\usage{
simplify.collect.log.exp.rules
}
\description{
reverse direction of \code{simplify.2} for \code{log} and \code{exp} expressions
}
\keyword{datasets}
