% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{simplify.pretty.01}
\alias{simplify.pretty.01}
\title{simplify.pretty.01}
\usage{
simplify.pretty.01(e)
}
\arguments{
\item{e}{expression}
}
\value{
simplied expression
}
\description{
translate \code{CONS('-',1)} to atomic -1
\code{ X^Y^-1 } to \code{ X/Y }
\code{ -X * Y } to \code{ -(X*Y) } which can be print as \code{-X*Y} using \code{\link{rexpression.to.string.nonmem}}
}
