% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{simplify.product.list}
\alias{simplify.product.list}
\title{simplify.product.list}
\usage{
simplify.product.list(le)
}
\arguments{
\item{le}{list of expression}
}
\value{
list of expression
}
\description{
Do following light checking to remove redundant terms in list of production \cr
remove 1's from list, if any entry equals to 0 , return 0 directly \cr
if \code{A} and \code{A^(-1)} both exists, remove both.
}
\author{
jjxie
}
