% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{simplify}
\alias{simplify}
\title{Simplify simplify an expression}
\usage{
simplify(e)
}
\arguments{
\item{e}{R expression}
}
\value{
simplified expression
}
\description{
Very elementary mathematical simplification of an R (symbolic) expression
\itemize{
 \item Arithmetic    \code{0*expression} \samp{->} \code{0}, \code{1*expression} \samp{->} \code{expression}, e.t.c
 \item  Logic        \code{expression && F} \samp{->} \code{F},  \code{expression || T} \samp{->} \code{T},   e.t.c        
 \item Other         \code{\{one expression\}} \samp{->} \code{one expression}, \code{(one expression)} \samp{->} \code{one expression}
 \item Atomic if both operands are atomic, then do the evaluation and replace them by the result
}
}
\author{
jjxie
}
