% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{simplify.sqrt}
\alias{simplify.sqrt}
\title{simplify.sqrt}
\usage{
simplify.sqrt(e)
}
\arguments{
\item{e}{expression}
}
\value{
simplified expression
}
\description{
simplify expression with \code{sqrt}
}
\author{
jjxie
}
