% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{simplify.sum.list}
\alias{simplify.sum.list}
\title{simplify.sum.list}
\usage{
simplify.sum.list(le)
}
\arguments{
\item{le}{list of expression}
}
\value{
simplified list of expression
}
\description{
Similar to \code{\link{simplify.product.list}}, do following light checking to remove redundant terms in list of summation \cr
remove 0's from list \cr
if \code{A} and \code{-A} both exists, remove both
}
\author{
jjxie
}
