% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.000common.R
\name{socp.filter.default}
\alias{socp.filter.default}
\title{default filter for set of critical points}
\usage{
socp.filter.default(gg, variables = NULL, eqns = NULL)
}
\arguments{
\item{gg}{graph object created by \code{\link{create.abstractGraph.from.equations}}}

\item{variables}{variables to be filtered}

\item{eqns}{definition equations}
}
\value{
the indices of socp for \code{variables}
}
\description{
People may write their own filter on purpose. However, this filtered is only used for picking up those \cr
known PK parameter (by \code{\link{guess.variable.meaning.by.name.default}}) which are also left hand side variables\cr
}
