% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10domain.000common.R
\name{socp.filter.known.pattern}
\alias{socp.filter.known.pattern}
\title{This filter uses the following idea: 
if a lhs variable is defined by some appearant pattern like lognormal or normal
It is probably being a critial variable}
\usage{
socp.filter.known.pattern(gg, variables = NULL, eqns = NULL)
}
\arguments{
\item{gg}{graph}

\item{variables}{variables list}

\item{eqns}{definition equations}
}
\value{
the indices of variables
}
\description{
This filter uses the following idea: 
if a lhs variable is defined by some appearant pattern like lognormal or normal
It is probably being a critial variable
}
