% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{sort.by.object.size}
\alias{sort.by.object.size}
\title{sort.by.object.size}
\usage{
\method{sort}{by.object.size}(le)
}
\value{
a list of terms
}
\description{
Internal use \cr
Sort a list of expressions by first, the size; then, deparsed result \cr
}
\author{
jjxie
}
