% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{specialize.estimable.unit}
\alias{specialize.estimable.unit}
\title{Specialize a estimable unit}
\usage{
specialize.estimable.unit(m, val = get.initial.estimation.estimable.unit(m),
  ind = seq_along(as.symbolic.vector(val)), simplify = FALSE)
}
\arguments{
\item{m}{estimable.unit}

\item{val}{specialize on which argument}

\item{ind}{optional index}

\item{simplify}{if simplify to simple atomic or expression, only possible give full value for vals}
}
\value{
new estimable.unit
}
\description{
Change estimable.unit to expression or simpler estimable.unit by specializing parameter values
or just substitute the initial estimations into the morphism or random disturbation.
}
