% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{specialize.estimable.units.estimable.model}
\alias{specialize.estimable.units.estimable.model}
\title{specialize \code{estimable.unit} into simpler one}
\usage{
specialize.estimable.units.estimable.model(esm, varnames, all = FALSE)
}
\arguments{
\item{esm}{estimable.model}

\item{varnames}{which variable need to be specialize}

\item{all}{boolean specialize all estimable.unit to their initial guess}
}
\value{
updated model
}
\description{
The background of this design is following:
NONMEM has convenient \code{FIX} notation for remove one parameter from to be estimabled.
Our infrastructure does not allow this, but specialization plays a same meaning.
Specilize some parameters to a given value means degenerate the original model into simpler one(with less parameters).
}
