% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{state.vector.nonmem.models.block}
\alias{state.vector.nonmem.models.block}
\title{rendering an ode or state.vector of ode to nonmem model description}
\usage{
state.vector.nonmem.models.block(stvs)
}
\arguments{
\item{stvs}{state.vectors or ode}
}
\value{
a string
}
\description{
rendering an ode or state.vector of ode to nonmem model description
}
