% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20pattern.05ode.R
\name{state.vectors}
\alias{state.vectors}
\title{generate state vectors
the storage of ode , the names and time variable are separated
need to combine them together}
\usage{
state.vectors(ode)
}
\arguments{
\item{ode}{ode system}
}
\value{
the vector function as state function
}
\description{
generate state vectors
the storage of ode , the names and time variable are separated
need to combine them together
}
