% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{summary.estimable.model}
\alias{summary.estimable.model}
\title{summary one estimable model in a data.frame}
\usage{
\method{summary}{estimable.model}(esm, all = FALSE)
}
\arguments{
\item{esm}{estimable.model}

\item{all}{if all variable are output}
}
\value{
data.frame
}
\description{
By default, the result returns only those information for basic PK parameters.
The option \code{all} can switch on the output for all variable appearing in the \code{PK.PROGRAM} and \code{OBS.PROGRAM}
}
