% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{summary.estimable.unit}
\alias{summary.estimable.unit}
\title{description for a estimable unit}
\usage{
\method{summary}{estimable.unit}(m)
}
\arguments{
\item{m}{estimable unit}
}
\value{
data.frame
}
\description{
similar to print, but results a data.frame
if input is not an estimable unit, return 0-rows data.frame
}
