% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.affine.mapping}
\alias{symbolic.affine.mapping}
\title{compute \code{Ax + b}}
\usage{
symbolic.affine.mapping(A, x, b = NULL)
}
\arguments{
\item{A}{symbolic matrix}

\item{x}{symbolic vector}

\item{b}{right vector}
}
\value{
result vector
}
\description{
compute \code{Ax + b}
}
