% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{symbolic.check.raw.equal}
\alias{symbolic.check.raw.equal}
\title{avoid the warnings if length is different}
\usage{
symbolic.check.raw.equal(e1, e2)
}
\arguments{
\item{e1}{symbol or call}

\item{e2}{symbol or call}
}
\value{
logical
}
\description{
avoid the warnings if length is different
}
