% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.CMatrix}
\alias{symbolic.CMatrix}
\title{similar to R matrix, columns first}
\usage{
symbolic.CMatrix(le, nrow = length(le), ncol = 1, byrow = FALSE)
}
\arguments{
\item{le}{list of expressions}

\item{nrow}{row number}

\item{ncol}{column number}

\item{byrow}{by row or by column}
}
\value{
list of list representation
}
\description{
similar to R matrix, columns first
}
