% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.010.ascii.art.R
\name{symbolic.draw.tree}
\alias{symbolic.draw.tree}
\title{display an expression as a tree}
\usage{
symbolic.draw.tree(e0, display = TRUE, do.not.expand = c("?", "?a", "?s",
  "?c", "?()", "?[]", "??", "?na"))
}
\arguments{
\item{e0}{expression}

\item{display}{\code{TRUE} for print the result out}

\item{do.not.expand}{We may avoid expanding the call starting by '?'. (If you do not want to expand '[', this can be set here)}
}
\value{
string for the tree
}
\description{
Only for debug and learning purpose
}
