% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{symbolic.grep}
\alias{symbolic.grep}
\title{symbolic.grep}
\usage{
symbolic.grep(pat, e)
}
\arguments{
\item{pat}{pattern expression}

\item{e}{expression from which to find the pattern}
}
\value{
\code{TRUE} or \code{FALSE} indicating if the pattern is found
}
\description{
recursively find a pattern in an expression
}
\author{
jjxie
}
