% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{symbolic.gsub}
\alias{symbolic.gsub}
\title{symbolic.gsub}
\usage{
symbolic.gsub(pat, rep, e0)
}
\arguments{
\item{pat}{pattern expression}

\item{rep}{expression as a replacement}

\item{e0}{expression to be operated}
}
\value{
the replaced expression
}
\description{
recursively replace a pattern \code{pat} in an expression \code{e0} by an expression \code{rep} \cr
  similar to \code{\link{symbolic.sub}} with exception that all the occurance of \code{pat} are replaced
}
\author{
jjxie
}
