% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.inner.product.logical}
\alias{symbolic.inner.product.logical}
\title{use \code{\link{symbolic.inner.product}} we get \code{(A==1)*a1 + (A==2)*a2}
however, use this function, we get \code{ifelse(A==1,a1,ifelse(A==2,a2,NA))}}
\usage{
symbolic.inner.product.logical(le1, le2)
}
\arguments{
\item{le1}{list of logical expression, as a symbolic vector}

\item{le2}{list of expression, as a symbolic vector}
}
\value{
one ifelse expression which is the inner product
}
\description{
use \code{\link{symbolic.inner.product}} we get \code{(A==1)*a1 + (A==2)*a2}
however, use this function, we get \code{ifelse(A==1,a1,ifelse(A==2,a2,NA))}
}
