% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.inner.product}
\alias{symbolic.inner.product}
\title{inner product for symbolic vectors}
\usage{
symbolic.inner.product(le1 = vectorize.symbol.and.atomic(le1),
  le2 = vectorize.symbol.and.atomic(le2))
}
\arguments{
\item{le1}{list of expression, as a symbolic vector}

\item{le2}{list of expression, as a symbolic vector}
}
\value{
one expression which is the inner product
}
\description{
inner product for symbolic vectors
}
