% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.listoflist.to.RMatrix}
\alias{symbolic.listoflist.to.RMatrix}
\title{convert between listoflist and symbolic.RMatrix}
\usage{
symbolic.listoflist.to.RMatrix(A)
}
\arguments{
\item{A}{a list of list}
}
\value{
RMatrix
}
\description{
convert between listoflist and symbolic.RMatrix
}
