% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{symbolic.match}
\alias{symbolic.match}
\title{symbolic.match}
\usage{
symbolic.match(pat, e0)
}
\arguments{
\item{pat}{pattern expression}

\item{e0}{expression to be matched against}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
A more handy interface for \code{\link{symbolic.match0}}. \cr
This funciton's return value is only \code{T/F}, but it has side effect on parent environment
}
\details{
e.g. for the example in symbolic.match0, it assigns back to the caller's environment  \cr
       \code{A = TVKA} \cr
       \code{B = 2}
}
\author{
jjxie
}
