% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.matrix.col.i}
\alias{symbolic.matrix.col.i}
\title{get column i}
\usage{
symbolic.matrix.col.i(A, i)
}
\arguments{
\item{A}{symbolic matrix (list of list)}

\item{i}{column index}
}
\value{
list as column
}
\description{
get column i
}
