% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.Matrix.mul}
\alias{symbolic.Matrix.mul}
\title{multiply two symbolic Matrix}
\usage{
symbolic.Matrix.mul(A, B)
}
\arguments{
\item{A}{Matrix}

\item{B}{Matrix}
}
\value{
RMatrix
}
\description{
multiply two symbolic Matrix
}
