% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.matrix.sub}
\alias{symbolic.matrix.sub}
\title{sub matrix defined by i,j}
\usage{
symbolic.matrix.sub(A, i, j)
}
\arguments{
\item{A}{matrix}

\item{i}{row index}

\item{j}{col index}
}
\value{
sub matrix
}
\description{
sub matrix defined by i,j
}
