% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.004ifelse.R
\name{symbolic.replace.rhss}
\alias{symbolic.replace.rhss}
\title{recursively replace right hand side
if (1) Y=a else Y=b
and try to replace b}
\usage{
symbolic.replace.rhss(pat, rep, e0)
}
\arguments{
\item{pat}{pattern}

\item{rep}{replacement}

\item{e0}{expression to be replaced}
}
\value{
new expression
}
\description{
recursively replace right hand side
if (1) Y=a else Y=b
and try to replace b
}
