% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.RMatrix.mul.RMatrix}
\alias{symbolic.RMatrix.mul.RMatrix}
\title{multiply two symbolic RMatrix}
\usage{
symbolic.RMatrix.mul.RMatrix(A, B)
}
\arguments{
\item{A}{RMatrix}

\item{B}{RMatrix}
}
\value{
RMatrix
}
\description{
multiply two symbolic RMatrix
}
