% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.RMatrix}
\alias{symbolic.RMatrix}
\title{symbolic list of list matrix}
\usage{
symbolic.RMatrix(le, nrow = length(le), ncol = 1, byrow = TRUE)
}
\arguments{
\item{le}{list of expressions}

\item{nrow}{row number}

\item{ncol}{column number}

\item{byrow}{by row or by column}
}
\value{
list of list representation
}
\description{
symbolic list of list matrix
}
