% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{symbolic.sub}
\alias{symbolic.sub}
\title{symbolic.sub}
\usage{
symbolic.sub(pat, rep, e0)
}
\arguments{
\item{pat}{pattern expression}

\item{rep}{expression as a replacement}

\item{e0}{expression to be operated}
}
\value{
the replaced expression
}
\description{
recursively replace a pattern \code{pat} in an expression \code{e0} by \code{rep} expression \cr
  Only the first occurance is replaced, \cr
  If the pattern is not found, \code{e0} is return without changing
}
\author{
jjxie
}
