% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{symbolic.symbol.extract}
\alias{symbolic.symbol.extract}
\title{symbolic.symbol.extract}
\usage{
symbolic.symbol.extract(e0)
}
\arguments{
\item{e0}{expression to be looked at}
}
\value{
a character vector containing unique symbols appearing in the expression
}
\description{
extract only symbols, excluding function call name and array names
}
\author{
jjxie
}
