% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.transpose.CMatrix}
\alias{symbolic.transpose.CMatrix}
\title{transpose CMatrix}
\usage{
symbolic.transpose.CMatrix(A)
}
\arguments{
\item{A}{CMatrix}
}
\value{
CMatrix
}
\description{
transpose CMatrix
}
