% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.transpose.RMatrix}
\alias{symbolic.transpose.RMatrix}
\title{transpose of RMatrix}
\usage{
symbolic.transpose.RMatrix(A)
}
\arguments{
\item{A}{RMatrix}
}
\value{
RMatrix
}
\description{
transpose of RMatrix
}
