% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{symbolic.vector.addition}
\alias{symbolic.vector.addition}
\title{symbolic Vector addition}
\usage{
symbolic.vector.addition(le1 = vectorize.symbol.and.atomic(le1),
  le2 = vectorize.symbol.and.atomic(le2))
}
\arguments{
\item{le1}{list of expression, as a symbolic vector}

\item{le2}{list of expression, as a symbolic vector}
}
\value{
one list which is the point-wise addition
}
\description{
symbolic Vector addition
}
