% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.003estimable.model.R
\name{table.random.parts.list.code}
\alias{table.random.parts.list.code}
\title{For a list of expressions or estimable units or random disturbation,
sketch a overall graph of the random parts}
\usage{
table.random.parts.list.code(lesu)
}
\arguments{
\item{lesu}{list of estimable.units or random disturbation}
}
\value{
number of random parts, start position in all, from new index to look up name, backref
}
\description{
If any random disturbation is found, 
a second \code{length}, \code{prestart}, \code{lookup} and \code{backref} are returned,
which are coded as \code{length2}, \code{prestart2}, \code{lookup2} and \code{backref2}.
}
\details{
Simple expression or names(symbols) are skipped. Only estimable unit and random disturbations are counted.
}
