% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.transformations.names.R
\name{transformations.names}
\alias{transformations.names}
\title{Get transformed regressors names}
\usage{
transformations.names(regressors, transformations)
}
\arguments{
\item{regressors}{array of regressors of type string}

\item{transformations}{list of transformations}
}
\value{
array of transformed regressors
}
\description{
Get transformed regressors names
}
