% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.000general.R
\name{union.symbolic}
\alias{union.symbolic}
\title{union.symbolic}
\usage{
union.symbolic(le = list(), x)
}
\arguments{
\item{le}{list of expressions}

\item{x}{expression}
}
\value{
list of unique expressons
}
\description{
union of symbolic expressions, similar to R's \code{union} for characters
}
