% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{update.all.initial.estimation.estimable.unit.in.estimable.model}
\alias{update.all.initial.estimation.estimable.unit.in.estimable.model}
\title{update the initial.estimations for all estimable unit}
\usage{
\method{update}{all.initial.estimation.estimable.unit.in.estimable.model}(esm,
  thetas)
}
\arguments{
\item{esm}{estimable.model}

\item{thetas}{a vector}
}
\value{
updated estimable model
}
\description{
For NONMEM users, it's common to consider all the parameters need to be estimated in a whole, no matter if 
\code{THETA[1]} and \code{THETA[2]} are both for modeling the first PK parameter \code{CL}.
So, we provide this API, user can use a vector of values \code{thetas} and used sequentially as the initial estimation for
all \code{estimable.unit}s
}
