% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.005estimable.model.api.R
\name{update.cov.estimable.model}
\alias{update.cov.estimable.model}
\title{fill in the estimable unit or random observations}
\usage{
\method{update}{cov.estimable.model}(esm, cov., indexed.by = "subject")
}
\arguments{
\item{esm}{estimable.model}

\item{cov.}{covariance matrix}

\item{indexed.by}{can be 'subject' or 'observation'}
}
\value{
new estimable.model
}
\description{
All inner random parameters are views as big random vector. And the \code{cov.} is viewed as
the covariance of the vector to update the model.
}
