% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{update.cov.estimable.unit}
\alias{update.cov.estimable.unit}
\title{update the covariance matrix of a estimable.unit}
\usage{
\method{update}{cov.estimable.unit}(m, Cov)
}
\arguments{
\item{m}{estimable.unit}

\item{Cov}{Variance or Covariance}
}
\value{
updated estimable.unit
}
\description{
Similar to \code{\link{update.distribution.estimable.unit}}, but
more common user story is we only want to change the covariance.
}
\details{
If the previous estimable unit has no random effect, we simple add one multi-normal distribution from give covariance matrix.
}
