% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{update.cov.random.observation}
\alias{update.cov.random.observation}
\title{update the covariance of distribution for a random.observation object}
\usage{
\method{update}{cov.random.observation}(m, Cov)
}
\arguments{
\item{m}{random observation}

\item{Cov}{covariance matrix}
}
\value{
updated random observation
}
\description{
More common situation is only change the covariance, not the whole type.
}
