% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15domain.150disturbation.R
\name{update.distribution.covariance}
\alias{update.distribution.covariance}
\title{update}
\usage{
\method{update}{distribution.covariance}(d, new.cov)
}
\arguments{
\item{d}{distribution}

\item{new.cov}{new covariance matrix}
}
\value{
updated distribution
}
\description{
update
}
