% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{update.distribution.estimable.unit}
\alias{update.distribution.estimable.unit}
\title{change the distribution attribute in a estimable unit}
\usage{
\method{update}{distribution.estimable.unit}(m, distr)
}
\arguments{
\item{m}{estimable.unit}

\item{distr}{distribution, type and Var}
}
\value{
updated estimable.unit
}
\description{
change the distribution attribute in a estimable unit
}
