% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{update.distribution.random.observation}
\alias{update.distribution.random.observation}
\title{update the distribution for a random.observation object}
\usage{
\method{update}{distribution.random.observation}(m, distr)
}
\arguments{
\item{m}{random observation}

\item{distr}{distribution object or list of type and Var}
}
\value{
updated random observation
}
\description{
By design, we record expression with \code{EPS[i]} as random observation.
}
