% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{update.domain.estimable.unit}
\alias{update.domain.estimable.unit}
\title{update the fixed domain of a estimable.unit if there is any}
\usage{
\method{update}{domain.estimable.unit}(m, domain)
}
\arguments{
\item{m}{estimable unit}

\item{domain}{domain expression}
}
\value{
updated estimable unit
}
\description{
Estimable unit can be implemented by either morphism or random disturbation. 
But the word domain here always related to the fixed part. For morphism, random effect is just an attribute,
for random disturbation, domain means the domain of fixed parameters.
}
