% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/70adhoc.001estimable.unit.R
\name{update.initial.estimation.estimable.unit}
\alias{update.initial.estimation.estimable.unit}
\title{update the FIXed parameters for estimable unit}
\usage{
\method{update}{initial.estimation.estimable.unit}(m, new.val)
}
\arguments{
\item{m}{estimable unit}
}
\value{
updated estimable unit
}
\description{
This function will check validity of the input value.
}
