% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00symbolic.006Matrix.R
\name{vectorize.symbol.and.atomic}
\alias{vectorize.symbol.and.atomic}
\title{enforce expression as a list}
\usage{
vectorize.symbol.and.atomic(e)
}
\arguments{
\item{e}{expression or an atomic vector}
}
\value{
list of the above
}
\description{
1:10 => list(1,2,3,4,...,10)
e => list(e)
}
