% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normailze.R
\name{normalize}
\alias{normalize}
\title{Normalize a dataset}
\usage{
normalize(X.df, custom.mins)
}
\arguments{
\item{X.df}{The input dataset to be normalized}

\item{custom.mins}{A list of user-defined minimum values for dataset columns.}
}
\value{
A list with two values:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    The input dataset normalized by subtracting the mean and dividing by the standard deviation.

    A list with some column statistic: std.min, mean, sd
}\if{html}{\out{</div>}}
}
\description{
Normalize a dataset
}
