% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe.r.squared.formula.len.fitness.R
\name{pe.r.squared.formula.len.fitness}
\alias{pe.r.squared.formula.len.fitness}
\title{Default fitness function for \code{genetic.search}}
\usage{
pe.r.squared.formula.len.fitness(errs.m, max.formula.len)
}
\arguments{
\item{errs.m}{a 1-row data.frame with cross-validation results}

\item{max.formula.len}{the maximum length of the formula we are searching for}
}
\value{
A single double value, the bigger the value the better the formula.
}
\description{
Default fitness function for \code{genetic.search}
}
\examples{
\dontrun{
 best.finetuned <- genetic.search(
                     best.vars.l=list(
                        c('inv.mul.patch_hyd_2.patch_hyd_2',
                          'mul.patch_hyd_5.patch_pos_.'),
                          ...
                         ),
                     complete.X.df=X,
                     n.squares=1,
                     max.formula.len=4,
                     y=Y,
                     maxiter=1000,
                     fitness.fun=pe.r.squared.formula.len.fitness
                  )
}
}
